### 2025-05-27 #8290 @editorialbot
[PRE REVIEW]: Abil: A Python package for the interpolation of aquatic biogeochemical datasets
**Submitting author:** @nanophyto (<a href="http://orcid.org/0000-0003-3427-6921">Joost de Vries</a>)
**Repository:** https://github.com/nanophyto/Abil
**Branch with paper.md** (empty if default branch):
**Version:** v25.05.1
