### 2025-05-27 nanophyto/Abil#255 @nanophyto
Fix problematic DOIs in paper
See https://github.com/openjournals/joss-reviews/issues/8290#issuecomment-2912321346
