### 2025-05-27 ffa3b54
adds missing DOIs
