+++ paper/paper.bib
@@ -18,6 +18,7 @@
   volume={45},
   pages={5--32},
   year={2001},
+  doi={10.1023/A:1010933404324},
   publisher={Springer}
 }
 @article{devries:2024,
@@ -28,6 +29,7 @@
   number={1},
   pages={920},
   year={2024},
+  doi={10.1038/s41597-024-03724-z},
   publisher={Nature Publishing Group UK London}
 }
 @article{dyck:2010,
@@ -39,6 +41,7 @@ journal = {Journal of Bioeconomics},
 pages = {227--243},
 title = "{Economic impact of ocean fish populations in the global fishery}",
 year = 2010,
+doi = {10.1007/s10818-010-9088-3},
 urldate = {2025-03-03}
 }
 
@@ -50,6 +53,7 @@ journal = {Philosophical Transactions of the Royal Society A},
 volume = {381},
 title = "{Sparse observations induce large biases in estimates of the global ocean CO2 sink: an ocean model subsampling experiment}",
 year = 2023,
+doi = {10.1098/rsta.2022.0063},
 urldate = {2025-03-03}
 }
 
@@ -61,7 +65,7 @@ volume = {12},
 number = {9},
 pages = {1620-1633},
 keywords = {machine learning, model transferability, predictive modelling, Random Forest, remote sensing, spatial mapping, uncertainty},
-doi = {https://doi.org/10.1111/2041-210X.13650},
+doi = {10.1111/2041-210X.13650},
 url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.13650},
 eprint = {https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/2041-210X.13650},
 abstract = {Abstract Machine learning algorithms have become very popular for spatial mapping of the environment due to their ability to fit nonlinear and complex relationships. However, this ability comes with the disadvantage that they can only be applied to new data if these are similar to the training data. Since spatial mapping requires predictions to new geographic space which in many cases goes along with new predictor properties, a method to assess the area to which a prediction model can be reliably applied is required. Here, we suggest a methodology that delineates the ‘area of applicability’ (AOA) that we define as the area where we enabled the model to learn about relationships based on the training data, and where the estimated cross-validation performance holds. We first propose a ‘dissimilarity index’ (DI) that is based on the minimum distance to the training data in the multidimensional predictor space, with predictors being weighted by their respective importance in the model. The AOA is then derived by applying a threshold which is the (outlier-removed) maximum DI of the training data derived via cross-validation. We further use the relationship between the DI and the cross-validation performance to map the estimated performance of predictions. We illustrate the approach in a simulated case study chosen to mimic ecological realities and test the credibility by using a large set of simulated data. The simulation studies showed that the prediction error within the AOA is comparable to the cross-validation error of the trained model, while the cross-validation error does not apply outside the AOA. This applies to models being trained with randomly distributed training data, as well as when training data are clustered in space and where spatial cross-validation is applied. Using the relationship between DI and cross-validation performance showed potential to limit predictions to the area where a user-defined performance applies. We suggest to add the AOA computation to the modeller's standard toolkit and to present predictions for the AOA only. We further suggest to report a map of DI-dependent performance estimates alongside prediction maps and complementary to (cross-)validation performance measures and the common uncertainty estimates.},
@@ -179,7 +183,8 @@ year = {2021}
   journal={Journal of Machine Learning Research},
   volume={12},
   pages={2825--2830},
-  year={2011}
+  year={2011},
[…]
