### 2025-05-27 de179b5
remove problematic DOI
