+++ paper/paper.bib
@@ -183,8 +183,7 @@ year = {2021}
   journal={Journal of Machine Learning Research},
   volume={12},
   pages={2825--2830},
-  year={2011},
-  doi={10.5281/zenodo.591564.}
+  year={2011}
 }
 
 @inproceedings{Chen:2016,
