### 2025-05-30 515fda0
fix "de Vries" bibtex formatting
