+++ paper/paper.bib
@@ -23,7 +23,7 @@
 }
 @article{devries:2024,
   title={CASCADE: Dataset of extant coccolithophore size, carbon content and global distribution},
-  author={de Vries, Joost and Poulton, Alex J and Young, Jeremy R and Monteiro, Fanny M and Sheward, Rosie M and Johnson, Roberta and Hagino, Kyoko and Ziveri, Patrizia and Wolf, Levi J},
+  author={{de Vries}, Joost and Poulton, Alex J and Young, Jeremy R and Monteiro, Fanny M and Sheward, Rosie M and Johnson, Roberta and Hagino, Kyoko and Ziveri, Patrizia and Wolf, Levi J},
   journal={Scientific data},
   volume={11},
   number={1},
