### 2025-06-17 6e25e05
Add optional argument to drop zeros before aoa
