### 2025-07-07 cb13643
add indexing check to avoid introducing NAs in missing X rows
