### 2025-07-07 f812fde
update indexing subset before AOA
