### 2025-07-08 4ca5e76
drop NAs before AOA
