### 2025-07-09 a89dddd
add some debugging print statements
