+++ abil/post.py
@@ -849,24 +849,24 @@ class post:
                 if isinstance(self.y_train, pd.Series):
                     y_col = self.y_train.name
                 else:  # DataFrame
-                    y_col = self.y_train.columns[0]  # assuming single column
+                    y_col = self.y_train.columns
                 
                 print("X_train is: ", print(self.X_train.head()))
                 print("y_train is: ", print(self.y_train.head()))
 
-                # 1. Merge X_train and y_train into one DataFrame
+                # Merge X_train and y_train into one DataFrame
                 combined = pd.concat([self.X_train, self.y_train], axis=1)
                 
                 print("combined head: ", combined.head())
-                combined = combined.dropna()
+                combined = combined.dropna(subset=self.X_train.columns)
                 print("length of combined: ", str(len(combined)))
-                # 2. Drop rows where y_train == 0
+                # Drop rows where y_train == 0
                 combined = combined[combined[y_col] > 0]
                 print("length of combined after dropping zeros: ", str(len(combined)))
 
-                # 3. Split back into X_train and y_train
-                X_train = combined.drop(columns=[y_col])
-                y_train = combined[y_col]
+                # Split back into X_train and y_train
+                X_train = combined[self.X_train.columns]
+                y_train = combined[combined.columns.difference(self.X_train.columns)]
                 
             else:
                 X_train = self.X_train
