### 2025-07-15 a37c468
use reg scoring for MAE estimates used for weights
