### 2025-07-15 b4d99c8
use reg instead of zir for R2, MAE and RMSE estimates
