### 2025-07-15 10fcdcf
export reg instead of zir R2 for posts
