### 2025-07-21 f517d35
Update post.py
Add options to specify threshold metric and return_all bool to access options within analyze.py
