### 2025-07-21 0bc4dac
Update post.py
allow return_all to be passed as true to area_of_applicability().
Add di_test output to aoa_dataset as {variable}_di.
Rename aoa variable to {variable}_aoa for clarity.
