### 2025-07-22 14d9f15
Update post.py
Add loop for return_all True or False that appends additional outputs to aoa dataframe. Also adds target_{output} to variables within aoa to specify output. Adds encoding datatype and compression for netcdf output.
