### 2025-07-22 17c7c7b
Update post.py
add encoding for return_all=False
