### 2025-07-24 952dd74
add implementation of nan-filtration in the aoa calculation
