### 2025-07-24 a4907aa
add nantest for the script section
