### 2025-08-01 #8290 @ruiying-ocean
> 👋 [@ruiying-ocean](https://github.com/ruiying-ocean) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @crvernon, thank you for the invitation to review this submission and congratuation to @nanophyto for finishing this package! Yes, I'd like to help, but I need to raise a potential conflict of interest - Joost and I were in the same research group before. Could you let me know if this would violate the JOSS's conflict of interest policies?
