### 2025-08-04 4e14ea6
downgrade checkarray syntax
