### 2025-08-04 d2b6f09
fix numpy.na call
