### 2025-08-04 1c623da
add drop zeros argument to aoa in post
