### 2025-08-05 #8290 @crvernon
Ok ,on those terms I think we pass the COI criteria and are good to proceed.  I'll find one more reviewer and then kick this off to full review.
