### 2025-08-12 nanophyto/Abil#267 @althonos
[JOSS] Add requirements as project dependencies
*This issue is part of JOSS review:* https://github.com/openjournals/joss-reviews/issues/8755

Hi! I've been assigned as a reviewer for your JOSS paper, and I will be listing my comments in the form of issues as I encounter them.

Currently, the package dependencies are not listed in the `pyproject.toml` project metadata, meaning that someone who does `pip install abil` won't have all the required packages for it to work. This is documented in the PyPA packaging guide: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-and-requirements.
