### 2025-08-13 nanophyto/Abil#268 @althonos
[JOSS] Consider providing a JSON/YAML schema and/or type hints
*This issue is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8755*

`abil` provides an API to handle reproducible models through a configuration file that can be used to define how models are ran. The YAML configuration format is described in the documentation.  The model configuration file is loaded into a Python dictionary using the `yaml` library, and the dictionary can be passed to the different classes of the `abil` package.

## Schema

At the moment, the only way to know the expected content of a file is to look at the documentation and/or copy-and-paste from the example file, which may be error-prone after modifications. To improve the distribution (and possible future compatibility) of model configuration files I suggest the following:

- Have a [JSON](https://json-schema.org/) or [YAML](https://www.asdf-format.org/projects/asdf-standard/en/1.0.1/schemas/yaml_schema.html) schema which describes the configuration file.
- Use the schema to validate data loaded from a YAML file (and thus having a dedicated `abil.load` function) rather than accepting any dictionary obtained from `yaml.load`: for instance, the [`yamale`](https://pypi.org/project/yamale/) package can be used to validate YAML with a YAML schema, and since it also uses `pyyaml` that wouldn't be too many new dependencies to add that functionality.
- Distribute the schema with the Python package, so that each `abil` release has an associated schema, making it easier to view changes in the schema is further options are added, renamed, or removed in the future.

## Python API

The section above addresses the configuration file, but the Python API could also be improved in the same direction to facilitate the creation of the `model_config` parameter expected by several functions and classes.

For instance, using a [typed dictionary](https://typing.python.org/en/latest/spec/typeddict.html) ([PEP 589](https://peps.python.org/pep-0589/)) can communicate to a static analyzer or an IDE which keys are expected in the `model_config` parameter of each function, thus making is easier to communicate which arguments are expected:

```python
import typing

class ModelConfig(typing.TypedDict):
    seed: int
    verbose: int
    path_out: Union[os.PathLike[str], str]
    ... # etc.

class predict:
    def __init__(self, X_train: pandas.DataFrame, y: pandas.Series, X_predict: pandas.DataFrame, model_config: ModelConfig):
        ....

```

Another option is to actually implement parameter objects which can be passed around, which would even further support static analysis. This way, serialization and deserialization from and to YAML could be handled with object methods, and validation could be made before creating the object. For instance using [`dataclasses`](https://docs.python.org/3/library/dataclasses.html):

```python
import dataclasses

@dataclasses.dataclass(frozen=True)
class ModelConfig:
    seed: int = 0
    verbose: int = 0
    path_out: Union[os.PathLike[str], str]

    @classmethod
    def load(cls, file):
        cfg = yaml.load(file)
        # add schema validation code here
        return cls( seed = cfg["seed"], verbose = cfg["verbose"], ...)
```
