### 2025-08-13 nanophyto/Abil#269 @althonos
[JOSS] Document tests and how to run them
*This issue is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8755*

The package ships with automated tests and has GitHub Actions configured to execute them on every commit, but the way to run the tests is not documented. Consider adding a *How to run tests* section to the *How to contribute* page of the documentation **given that the standard `python -m unittest discover` is not sufficient to execute the tests with your setup**.

Also consider adding a sentence or two to the paper to document that you're running automated tests and continuous integration through GitHub CI, as this is a good feature that improves the chances of a reader using your software :+1:
