### 2025-08-13 nanophyto/Abil#270 @althonos
[JOSS] Implement tests that actually check for produced results
*This issue is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8755*

This project implement tests written with the `unittest` framework of the standard library, however at the moment the tests merely consist in executing code adapted from the `2-phase.py` and `regressor.py` examples and "checking it doesn't crash". At no point do the test actually validate the functionalities of code, or the values predicted by the models, which should be reproducible.

Rather than using generated data, I would encourage including a small but real dataset (or using the ones from the examples), and in the unit tests verifying that the values predicted by the model with the given model configuration are always the same. This would allow running the tests on different machines, with possibly different OS or architectures, while ensuring a pair of model configuration and training dataset always produce consistent results independently of the execution context.
