### 2025-08-13 nanophyto/Abil#271 @althonos
[JOSS] Avoid Python anti-patterns
*This issue is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8755*

The code contains several Python anti-patterns that should be addressed to improve usability and flexibility of the API.

## Avoid using assertions to check values

As assertions can always be disabled using the `python -O` (optimize) flag, they should not be used for data validation, e.g. in `analyze.py`:

```python
assert n_features == X_train.shape[1]("features must be the same for both training and test data.")
```

should be turned into proper exceptions:

```python
if n_features != X_train.shape[1]:
    raise ValueError("features must be the same for both training and test data.")
```

## Avoid printing in library code

The code is currently printing a lot of information to `stdout`, which shouldn't be the case for a library. To communicate information I suggest using the [`logging`](https://docs.python.org/3/library/logging.html) module of the standard library, which has the advantage that it can be turned off by consumer code if needed. You could then display logger messages when executing `abil` code from a script, but keep it silent otherwise.

## Avoid setting environment values in library files

In `tune.py`:

```python
os.environ['MKL_NUM_THREADS'] = '1'
os.environ['OMP_NUM_THREADS'] = '1'
os.environ['MKL_DYNAMIC'] = 'FALSE'
os.environ["OMP_THREAD_LIMIT"] = "1"
```

These lines appear at the top of the module, meaning that simply running `import abil.tune` will cause these variables to be set and possibly modify the behaviour of the importing program, which is really not desirable. Ensure that the file have no side-effect, and move this to the script section (under the `if __name__ == "__main__"`).

## [PEP-8](https://peps.python.org/pep-0008) compliance

### Class names

> Class names should normally use the CapWords convention.

*NB: I would also recommend consistency between the class name and its main method, e.g. renaming `Tune` to `Trainer` and `Predict` to `Predictor`, so that the "main" method of each classes are `Trainer.train` and `Predictor.predict` respectively.*

### Boolean value comparison

> Don’t compare boolean values to `True` or `False` using `==`

E.g. in `predict.py`:

```python
if (self.ensemble_config["classifier"] ==True) and (self.ensemble_config["regressor"] == False):
```
should be replaced with:
```python
if self.ensemble_config["classifier"] and not self.ensemble_config["regressor"]:
```

### Unbound `try/except` blocks

> When catching exceptions, mention specific exceptions whenever possible instead of using a bare `except` clause

E.g. in `tune.py`:

```python
try:
    print("reg rRMSE: " + str(int(round(np.mean(reg_scores['test_RMSE'])/np.mean(self.y), 2)*-100))+"%")
except:
    print("reg rRMSE is NA (!)")
```

## Directory creation

E.g. in `tune.py`:

```python
try: #make new dir if needed
    os.makedirs(reg_sav_out_scores)
except:
    None
```

Replace with:

```python
os.makedirs(reg_sav_out_scores, exist_ok=True)
```

This will create the directory if it doesn't exist, but avoid silencing other exceptions that may be worth reporting to the user, for instance if creating the directory failed because of permission errors or invalid characters in the path name.
