### 2025-08-13 nanophyto/Abil#272 @althonos
[JOSS] Consider writing an example with more context
*This issue is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8755*

The two examples in the API show different use-cases for Abil, but I think there could be some more value in having an example that actually goes a bit more into context of what is being achieved. The problem of using the `example_data` is that:

- for someone not familiar with how the data looks like, there is not a single point in the guide where we actually show the example data, so the rest of the guide is very abstract in terms of what is being expected to train or produce results
- for someone who wants to make there own

I would suggest the following:

- Use an example with executable code in the form of a Jupyter notebook (which can be easily integrated into the documentation using [`nbsphinx`](https://nbsphinx.readthedocs.io/en/0.9.7/)), so that the commands are actually executed and the user can see the results
- Use an example after a real situation, not from the `example_data` function, so that it's clearer what the data looks like and what question the code can be used to answer.

Ideally having an example notebook that produces something like Figure 1 of the paper would be extremely valuable, compared to the current examples which remain somewhat abstract for someone outside the field.
