### 2025-08-16 3ee659f
adds Zenodo badge
