+++ README.md
@@ -1,6 +1,6 @@
 
 
-# Abil.py &middot; [![GitHub license](https://img.shields.io/badge/license-MIT-blue.svg)](https://github.com/nanophyto/Abil/LICENSE) [![Build Status](https://github.com/nanophyto/Abil/actions/workflows/ci.yml/badge.svg?branch=Continuous-integration)](https://github.com/nanophyto/Abil/actions/workflows/ci.yml?query=branch%3AContinuous-integration) [![dev](https://img.shields.io/badge/docs-Dev_Docs-blue)](https://nanophyto.github.io/Abil/)
+# Abil.py &middot; [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.16886567.svg)](https://doi.org/10.5281/zenodo.16886567) [![GitHub license](https://img.shields.io/badge/license-MIT-blue.svg)](https://github.com/nanophyto/Abil/LICENSE) [![Build Status](https://github.com/nanophyto/Abil/actions/workflows/ci.yml/badge.svg?branch=Continuous-integration)](https://github.com/nanophyto/Abil/actions/workflows/ci.yml?query=branch%3AContinuous-integration) [![dev](https://img.shields.io/badge/docs-Dev_Docs-blue)](https://nanophyto.github.io/Abil/)
 
 ## Overview
 
