### 2025-08-18 nanophyto/Abil#273 @ruiying-ocean
[JOSS] installation and example data
Review comment for https://github.com/openjournals/joss-reviews/issues/8755

## No dependencies declared

I installed the package using pip, but the dependencies were not downlowded automatically
These include:
pandas, joblib, xgboost xarray, scikit-bio, pyyaml (not yaml)

so you might want to add them into pyproject.toml: https://peps.python.org/pep-0631/

## Examples in the README

The instruction is in the documentation site, but the accessibility could be improved by adding some quick examples in the README, so people know how this project can be used

## Example csv data not stored in the package

When I followed the [instruction for one-phase ensemble](https://nanophyto.github.io/Abil/source/user-guide/one-phase-ensemble.html), it can't run successfully, with errors raised in *Post-processing* step:

This can be configured in the yaml file, but it'll be better if it can work as in the example directly.

```python
FileNotFoundError                         Traceback (most recent call last)
Cell In[7], line 19
     16     integ.integrated_totals(targets, monthly=True)
     17     integ.integrated_totals(targets)
---> 19 do_post(statistic="mean")
     20 do_post(statistic="ci95_UL")
     21 do_post(statistic="ci95_LL")

Cell In[7], line 3, in do_post(statistic)
      2 def do_post(statistic):
----> 3     m = post(X_train, y, X_predict, model_config, statistic, datatype="poc")
      5     m.estimate_applicability()
      6     m.estimate_carbon("pg poc")

File ~/miniforge3/envs/abil/lib/python3.13/site-packages/abil/post.py:120, in post.__init__(self, X_train, y_train, X_predict, model_config, statistic, datatype)
    118 self.path_out = os.path.join(model_config['root'], model_config['path_out'], model_config['run_name'], "posts/")
    119 self.ds = merge_netcdf(os.path.join(model_config['root'], model_config['path_out'], model_config['run_name'], "predictions", "ens"), statistic)
--> 120 self.traits = pd.read_csv(os.path.join(model_config['root'], model_config['targets']))
    122 self.root  =  model_config['root']
    123 self.statistic = statistic

File ~/miniforge3/envs/abil/lib/python3.13/site-packages/pandas/io/parsers/readers.py:1026, in read_csv(filepath_or_buffer, sep, delimiter, header, names, index_col, usecols, dtype, engine, converters, true_values, false_values, skipinitialspace, skiprows, skipfooter, nrows, na_values, keep_default_na, na_filter, verbose, skip_blank_lines, parse_dates, infer_datetime_format, keep_date_col, date_parser, date_format, dayfirst, cache_dates, iterator, chunksize, compression, thousands, decimal, lineterminator, quotechar, quoting, doublequote, escapechar, comment, encoding, encoding_errors, dialect, on_bad_lines, delim_whitespace, low_memory, memory_map, float_precision, storage_options, dtype_backend)
   1013 kwds_defaults = _refine_defaults_read(
   1014     dialect,
   1015     delimiter,
   (...)   1022     dtype_backend=dtype_backend,
   1023 )
   1024 kwds.update(kwds_defaults)
-> 1026 return _read(filepath_or_buffer, kwds)

File ~/miniforge3/envs/abil/lib/python3.13/site-packages/pandas/io/parsers/readers.py:620, in _read(filepath_or_buffer, kwds)
    617 _validate_names(kwds.get("names", None))
    619 # Create the parser.
--> 620 parser = TextFileReader(filepath_or_buffer, **kwds)
    622 if chunksize or iterator:
    623     return parser

File ~/miniforge3/envs/abil/lib/python3.13/site-packages/pandas/io/parsers/readers.py:1620, in TextFileReader.__init__(self, f, engine, **kwds)
   1617     self.options["has_index_names"] = kwds["has_index_names"]
   1619 self.handles: IOHandles | None = None
-> 1620 self._engine = self._make_engine(f, self.engine)

File ~/miniforge3/envs/abil/lib/python3.13/site-packages/pandas/io/parsers/readers.py:1880, in TextFileReader._make_engine(self, f, engine)
   1878     if "b" not in mode:
   1879         mode += "b"
-> 1880 self.handles = get_handle(
   1881     f,
   1882     mode,
   1883     encoding=self.options.get("encoding", None),
   1884     compression=self.options.get("compression", None),
   1885     memory_map=self.options.get("memory_map", False),
   1886     is_text=is_text,
   1887     errors=self.options.get("encoding_errors", "strict"),
   1888     storage_options=self.options.get("storage_options", None),
   1889 )
   1890 assert self.handles is not None
   1891 f = self.handles.handle

File ~/miniforge3/envs/abil/lib/python3.13/site-packages/pandas/io/common.py:873, in get_handle(path_or_buf, mode, encoding, compression, memory_map, is_text, errors, storage_options)
    868 elif isinstance(handle, str):
    869     # Check whether the filename is to be opened in binary mode.
    870     # Binary mode does not support 'encoding' and 'newline'.
    871     if ioargs.encoding and "b" not in ioargs.mode:
    872         # Encoding
--> 873         handle = open(
    874             handle,
    875             ioargs.mode,
    876             encoding=ioargs.encoding,
    877             errors=errors,
    878             newline="",
    879         )
    880     else:
    881         # Binary mode
    882         handle = open(handle, ioargs.mode)

FileNotFoundError: [Errno 2] No such file or directory: './docs/examples/data/targets.csv'
```
