### 2025-08-18 nanophyto/Abil#274 @ruiying-ocean
[JOSS] class name convention
review for https://github.com/openjournals/joss-reviews/issues/8755

I suggest using uppercase letter for all the class names, like `Post` rather than `post`, to follow the [PEP8 guidelines] (https://peps.python.org/pep-0008/)

Also, the current workflow is tune-predict-post. I suggest using nouns instead of verbs for class names, which are more clear. Some examples:

tune ➡️ ModelTuner
predict ➡️ Predictor/ModelPredictor
post ➡️ PostProcessor/PostAnalyser
