### 2025-08-22 nanophyto/Abil#275 @ruiying-ocean
[JOSS] reference check
This is a review comment for https://github.com/openjournals/joss-reviews/issues/8755

Some references in the paper can be improved:

## Scikit-learn doesn't have a doi

https://doi.org/10.48550/arXiv.1201.0490

## Author of Pandas reference is not rendered properly

@misc{pandas2020,
  author = {The pandas development team},
  title = {pandas-dev/pandas: Pandas},
  year = {2020},
  publisher = {Zenodo},
  journal = {GitHub repository},
  doi = {10.5281/zenodo.3509134},
  url = {https://doi.org/10.5281/zenodo.3509134}
}

## Same for joblib with citation.cff

@misc{joblib2025,
  author = {The joblib developers},
  title = {joblib},
  year = {2025},
  publisher = {Zenodo},
  journal = {GitHub repository},
  doi = {10.5281/zenodo.14915601},
  url = {https://github.com/joblib/joblib}
}
