### 2025-08-22 nanophyto/Abil#276 @ruiying-ocean
[JOSS] unit test failed
This is a review comment for https://github.com/openjournals/joss-reviews/issues/8755

## Unsuccessful automated test

I ran `python -m unittest tests/test.py`, but it throws an error:

``` python
ValueError: Cannot register /var/folders/1f/q0z6q0wx7h70hvl16srdbsrh5tw3wt/T/joblib_memmapping_folder_8752_47c8eafbb01f49ca9636f565e0c842f4_484e153c0af74ba78b3535c7635e3c2e for automatic cleanup: unknown resource type folder
Traceback (most recent call last):
  File "/Users/vhf24tbu/.local/share/uv/python/cpython-3.13.7-macos-aarch64-none/lib/python3.13/multiprocessing/resource_tracker.py", line 295, in main
    raise ValueError(
        f'Cannot register {name} for automatic cleanup: '
        f'unknown resource type {rtype}')
```

I might run it wrongly, but it's better to have an instruction in the README or your documentation site.
