### 2025-08-22 nanophyto/Abil#274 @ruiying-ocean
Just to say, this can be done using vs code's rename symbol function
