### 2025-09-01 059befa
expand example with real data
