+++ docs/examples/regressor.py
@@ -1,26 +1,38 @@
 """
 2-phase Ensemble 
 """
+#handling data:
 import numpy as np
+import pandas as pd
+import xarray as xr
 from yaml import load
 from yaml import CLoader as Loader
+#plotting:
+import matplotlib.pyplot as plt
+import cartopy.crs as ccrs
+import cartopy.feature as cfeature
+#abil functions:
 from abil.tune import tune
 from abil.predict import predict
 from abil.post import post
 from abil.utils import example_data 
-
+#paths:
 import os
-os.chdir('./docs/examples/')
+
+os.chdir(os.path.join(".", "docs", "examples"))
 
 #load configuration yaml:
 with open('regressor.yml', 'r') as f:
     model_config = load(f, Loader=Loader)
-
-#create some example training data:
-target_name =  "Emiliania huxleyi"
-X_train, X_predict, y = example_data(target_name, n_samples=1000, n_features=3, 
-                                    noise=0.1, train_to_predict_ratio=0.7, 
-                                    random_state=59)
+#load example training data:
+d = pd.read_csv(os.path.join("..", "data", "training.csv"))
+#define target:
+target = "Gephyrocapsa huxleyi HET"
+#define predictors based on YAML:
+predictors = model_config['predictors']
+#split training data in X_train and y
+y = d[target]
+X_train = d[predictors]
 
 #train your model:
 m = tune(X_train, y, model_config)
@@ -28,29 +40,100 @@ m.train(model="rf")
 m.train(model="xgb")
 m.train(model="knn")
 
+#load prediction data:
+X_predict = pd.read_csv(os.path.join("..", "data", "southern_ocean_surface.csv"))
+X_predict.set_index(['lat', 'lon'], inplace=True)
+
 #predict your model:
 m = predict(X_train, y, X_predict, model_config)
 m.make_prediction()
 
 # Posts
-targets = np.array([target_name])
+targets = np.array([target])
 def do_post(statistic):
     m = post(X_train, y, X_predict, model_config, statistic, datatype="poc")
-    
-    m.estimate_applicability()
     m.estimate_carbon("pg poc")
-    m.total()
-
-    m.merge_env()
-    m.merge_obs("predictions_obs", targets)
-
     m.export_ds("my_first_regressor_model")
-
-    vol_conversion = 1e3 #L-1 to m-3
-    integ = m.integration(m, vol_conversion=vol_conversion)
-    integ.integrated_totals(targets, monthly=True)
-    integ.integrated_totals(targets)
 
 do_post(statistic="mean")
 do_post(statistic="ci95_UL")
 do_post(statistic="ci95_LL")
+
+# Load the predictions
+ds = xr.open_dataset("./ModelOutput/regressor/posts/my_first_regressor_model_mean_poc.nc")
+ds_UL = xr.open_dataset("./ModelOutput/regressor/posts/my_first_regressor_model_ci95_UL_poc.nc")
+ds_LL = xr.open_dataset("./ModelOutput/regressor/posts/my_first_regressor_model_ci95_LL_poc.nc")
+
+# Log-transform data
+ds['Gephyrocapsa huxleyi HET'] = np.log10(ds['Gephyrocapsa huxleyi HET'] + 1)
[…]
