### 2025-09-01 6d7cfdb
Update example text to be more descriptive
