### 2025-09-01 nanophyto/Abil#276 @nicola-wiseman
Hi @ruiying-ocean ,
I've tried to replicate this error. I've just freshly downloaded the package and run the same command but had no errors pop up.
Could you confirm what versions of joblib and scikit-learn are being used within the environment?
Mine is currently working with joblib=1.4.0, scikit-learn=1.3.0
