### 2025-09-02 a4b18c6
remove asserts
