+++ abil/post.py
@@ -4,6 +4,7 @@ import glob, os
 import xarray as xr
 import pickle
 import gc
+import logging
 from yaml import dump, Dumper
 from skbio.diversity.alpha import shannon
 
@@ -76,7 +77,7 @@ class post:
 
             This function uses `xarray.open_mfdataset` to load all NetCDF files in the given directory 
             (matching the pattern "*.nc") and combines them into one xarray.Dataset. The function 
-            prints status messages indicating the start and completion of the merging process.
+            logging.info(s status messages indicating the start and completion of the merging process.
 
             Parameters
             ----------
@@ -91,8 +92,8 @@ class post:
                 The merged dataset containing the combined data from all the NetCDF files in the directory.
                 The variable names in the merged dataset are derived from the 'target' values in each file.
             """
-            print("merging...")
-            print(path_in)
+            logging.info("merging...")
+            logging.info(path_in)
 
             datasets = []
             
@@ -107,12 +108,12 @@ class post:
                         ds_subset = ds[[statistic]].rename({statistic: target_name})
                         datasets.append(ds_subset)
                     else:
-                        print(f"Statistic '{statistic}' not found in {file}")
+                        logging.info(f"Statistic '{statistic}' not found in {file}")
 
             # Merge datasets by variables, keeping same coordinates
             merged_ds = xr.merge(datasets, compat='override')  # 'override' skips conflicts
 
-            print("finished merging NetCDF files")
+            logging.info("finished merging NetCDF files")
             return merged_ds
 
         self.path_out = os.path.join(model_config['root'], model_config['path_out'], model_config['run_name'], "posts/")
@@ -162,7 +163,7 @@ class post:
         ------
         Exception
             If an error occurs during the directory creation or file writing process, an exception
-            is caught and an error message is printed.
+            is caught and an error message is logging.info(ed.
 
         Notes
         -----
@@ -176,9 +177,9 @@ class post:
             with open(yml_file_path, 'w') as yml_file:
                 dump(self.model_config, yml_file, Dumper=Dumper, default_flow_style=False)
             
-            print(f"Model configuration exported to: {yml_file_path}")
+            logging.info(f"Model configuration exported to: {yml_file_path}")
         except Exception as e:
-            print(f"Error exporting model_config to YAML: {e}")   
+            logging.info(f"Error exporting model_config to YAML: {e}")   
 
     def merge_performance(self):
         """
@@ -191,8 +192,8 @@ class post:
         """    
 
         models = [value for key, value in self.model_config['ensemble_config'].items() if key.startswith("m")]
-        print("models included in merge performance!")
-        print(models)
+        logging.info("models included in merge performance!")
+        logging.info(models)
         models.append("ens")
         for model in models:
             self.merge_performance_single_model(model)
@@ -262,7 +263,7 @@ class post:
             None
         all_performance.to_csv(os.path.join(self.root, self.model_config['path_out'], self.model_config['run_name'], "posts/performance", model) + "_performance.csv", index=False)
 
-        print("finished merging performance")
+        logging.info("finished merging performance")
 
     def merge_parameters(self):
         """
@@ -316,8 +317,8 @@ class post:
         for i in range(len(self.unique_targets)):
             
             target = self.unique_targets[i]
-            print("the target is:")
-            print(target)
+            logging.info("the target is:")
+            logging.info(target)
             target_no_space = target.replace(' ', '_')
 
[…]
