+++ abil/tune.py
@@ -8,6 +8,7 @@ import time
 import pickle
 import pandas as pd
 import numpy as np
+import logging
 from joblib import parallel_backend
 from xgboost import XGBClassifier, XGBRegressor
 from sklearn.model_selection import GridSearchCV, cross_validate, KFold
@@ -90,8 +91,7 @@ class tune:
         
         """
         self.y = y.sample(frac=1, random_state=model_config['seed']) #shuffle
-        print("length of y:")
-        print(len(self.y))
+        logging.info(f"length of y: {len(self.y)}")
         self.y = self.y.values.ravel()
         self.X_train = X_train.sample(frac=1, random_state=model_config['seed']) #shuffle
         self.model_config = model_config
@@ -113,7 +113,7 @@ class tune:
         if model_config['stratify']:
             if model_config['upsample']:
                 self.cv = UpsampledZeroStratifiedKFold(n_splits=model_config['cv'])
-                print("upsampling = True")
+                logging.info("upsampling = True")
             else:
                 self.cv = ZeroStratifiedKFold(n_splits=model_config['cv'])
         else:
@@ -217,7 +217,7 @@ class tune:
                 X_train = self.X_train
                 cv = self.cv
 
-            print("training regressor")
+            logging.info("training regressor")
 
             reg_scoring = {
                 'R2': 'r2',
@@ -233,7 +233,7 @@ class tune:
                 for key, value in user_reg_param_grid.items()
             }
 
-            print(reg_param_grid)
+            logging.info(reg_param_grid)
 
             reg_sav_out_scores = os.path.join(self.path_out, "scoring/", model)
             reg_sav_out_model = os.path.join(self.path_out, "model/", model)
@@ -263,7 +263,7 @@ class tune:
             with open(os.path.join(reg_sav_out_model, self.target_no_space) + '_reg.sav', 'wb') as f:
                 pickle.dump(m2, f)
 
-            print("exported model to: " + reg_sav_out_model + "/"  + self.target_no_space + '_reg.sav')
+            logging.info("exported model to: " + reg_sav_out_model + "/"  + self.target_no_space + '_reg.sav')
 
             with parallel_backend('multiprocessing', n_jobs=self.n_jobs):
                 reg_scores = cross_validate(m2, X_train, y, cv = cv, verbose = self.verbose, scoring=reg_scoring)
@@ -271,27 +271,27 @@ class tune:
             with open(os.path.join(reg_sav_out_scores, self.target_no_space) + '_reg.sav', 'wb') as f:
                 pickle.dump(reg_scores, f)
 
-            print("exported scoring to: " + reg_sav_out_scores + "/" + self.target_no_space + '_reg.sav')
+            logging.info("exported scoring to: " + reg_sav_out_scores + "/" + self.target_no_space + '_reg.sav')
 
             if "RMSE" in reg_scoring:
                 try:
-                    print("reg rRMSE: " + str(int(round(np.mean(reg_scores['test_RMSE'])/np.mean(self.y), 2)*-100))+"%")
+                    logging.info("reg rRMSE: " + str(int(round(np.mean(reg_scores['test_RMSE'])/np.mean(self.y), 2)*-100))+"%")
                 except:
-                    print("reg rRMSE is NA (!)")
+                    logging.info("reg rRMSE is NA (!)")
             if "MAE" in reg_scoring:
                 try:
-                    print("reg rMAE: " + str(int(round(np.mean(reg_scores['test_MAE'])/np.mean(self.y), 2)*-100))+"%")
+                    logging.info("reg rMAE: " + str(int(round(np.mean(reg_scores['test_MAE'])/np.mean(self.y), 2)*-100))+"%")
                 except:
[…]
