### 2025-09-02 8a075cc
move env var setting to main body
