### 2025-09-02 d05b964
swap to exist_ok
