### 2025-09-02 nanophyto/Abil#271 @ljwolf
To flag one thing: the behaviors of the **Directory Creation** code will be different than the existing code. If the existing code does not have write permissions to the target, it will skip writing. If the `os.makedirs()` call does not have write permissions, it will raise `PermissionError`. @nanophyto I'm assuming it's fine for this revision to raise the permission error early?
