### 2025-09-02 nanophyto/Abil#276 @ruiying-ocean
Hi @nicola-wiseman ,

This is my dependency list, which is also created from scratch.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "abil"
version = "25.08.06"
license = "MIT"
license-files = ['LICENSE']
description = "Aquatic Biogeochemical Interpolation Library"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "ipykernel>=6.16.2",
    "joblib>=1.3.2",
    "matplotlib>=3.5.3",
    "netcdf4>=1.6.5",
    "pandas>=1.1.5",
    "pytest>=7.4.4",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.0.2",
    "xarray>=0.20.2",
]

[project.urls]
"Homepage" = "https://github.com/nanophyto/Abil"

[tool.hatch.build]
exclude = [
  "/.git",
  "/studies",
  "/dist",
]

I run the following command to test
```bash
python -m unittest test.py
```
