### 2025-09-02 nanophyto/Abil#274 @ljwolf
Renamed `class Tune` to `class ModelTuner`, `class predict` to `class ModelPredictor`, and `class post` to `class SDMPostProcessor`. I chose the "SDM" prefix for the final `post` class because it contains quite a few things specific to SDM models, while the former could be applied to a more generic context.
