### 2025-09-02 36f8505
update code for symbol renames across examples and tests
