+++ tests/test.py
@@ -8,10 +8,10 @@ import pandas as pd
 import numpy as np
 import xarray as xr 
 
-from abil.tune import tune
+from abil.tune import ModelTuner as tune
 from abil.utils import example_data # example_training_data, example_predict_data
-from abil.predict import predict
-from abil.post import post
+from abil.predict import ModelPredictor as predict
+from abil.post import SDMPostProcessor as post
 
 class TestRegressors(unittest.TestCase):
 
