### 2025-09-03 #8755 @althonos
Sorry @crvernon !

I share the comments posted by @ruiying-ocean: to me this is a useful package, but there are some improvements that are needed to strengthen confidence in the code. I haven't tested the HPC part yet but I have access to a SLURM cluster so I can attempt that next. I have no comments on the paper which I think is well written and shows an interesting example on real data.

1. In terms of Python code, there are some updates that should be made to standardize the API regarding naming conventions ([nanophyto/Abil#271](https://github.com/nanophyto/Abil/issues/271)), avoiding to print in library code, fixing the packaging so that the dependencies are installed automatically ([nanophyto/Abil#267](https://github.com/nanophyto/Abil/issues/267)), etc.
2. In terms of scientific code, the package is not tested in a way that ensure reproducibility, which in my view is a blocker for a paper that claims reproducibility ([nanophyto/Abil#269](https://github.com/nanophyto/Abil/issues/269), [nanophyto/Abil#270](https://github.com/nanophyto/Abil/issues/270)). I think it is really imperative to have tests that ensure the results produced by `abil` on a given test case (not just executing the code but actually comparing the values to an expected reference) to really ensure that the models behave the same across platforms, `scikit-learn` versions, etc.
3. In terms of functionality and software engineering, I think that the package would greatly benefit from having validation of the configuration files through a schema ([nanophyto/Abil#268](https://github.com/nanophyto/Abil/issues/268)). This is pretty standard to do with JSON or YAML, and would really be helpful in catching configuration issues.

I am also happy to see the paper accepted after revision, I think point 1 and 2 *need* to be addressed, but I greatly encourage tackling point 3 as well, as I can see benefits in forward compatibility and overall user-experience.
