### 2025-09-03 9745430
update to model BATS
