+++ docs/examples/regressor.py
@@ -1,38 +1,55 @@
 """
 2-phase Ensemble 
 """
+#paths:
+import os
 #handling data:
 import numpy as np
 import pandas as pd
 import xarray as xr
 from yaml import load
 from yaml import CLoader as Loader
-#plotting:
-import matplotlib.pyplot as plt
-import cartopy.crs as ccrs
-import cartopy.feature as cfeature
 #abil functions:
 from abil.tune import tune
 from abil.predict import predict
 from abil.post import post
-from abil.utils import example_data 
-#paths:
-import os
+#plotting:
+import matplotlib.pyplot as plt
 
 os.chdir(os.path.join(".", "docs", "examples"))
 
 #load configuration yaml:
-with open('regressor.yml', 'r') as f:
+with open('regressor_BATS.yml', 'r') as f:
     model_config = load(f, Loader=Loader)
 #load example training data:
-d = pd.read_csv(os.path.join("..", "data", "training.csv"))
+d = pd.read_csv(os.path.join("..", "data", "bats_training.csv"))
 #define target:
-target = "Gephyrocapsa huxleyi HET"
+target= "Gephyrocapsa huxleyi"
 #define predictors based on YAML:
 predictors = model_config['predictors']
 #split training data in X_train and y
 y = d[target]
 X_train = d[predictors]
+
+#plot training data:
+#first we setup a helper function to plot the rasters:
+def plot_depth_time(ax, da, variable=None, vmin=None, vmax=None,
+             cbar_label=None):
+    if isinstance(da, pd.DataFrame):
+        da = da.to_xarray()
+    if isinstance(da, xr.Dataset):
+        da = da[variable]  
+    m=da.plot(ax=ax,add_colorbar=True,
+              vmin=vmin,vmax=vmax)
+    m.colorbar.set_label(cbar_label)
+    return ax
+# then we can make the plots
+fig, ax = plt.subplots(figsize=(14,5))
+d.set_index(['depth', 'time'], inplace=True) # for plotting
+plot_depth_time(ax, d, variable=target, cbar_label=r'abundance (cells L$^{-1}$)')
+ax.set_title("Gephyrocapsa huxleyi abundance at BATS")
+plt.savefig('BATS_observations.png',dpi=300,bbox_inches='tight',facecolor='white')
+plt.show()
 
 #train your model:
 m = tune(X_train, y, model_config)
@@ -41,8 +58,18 @@ m.train(model="xgb")
 m.train(model="knn")
 
 #load prediction data:
-X_predict = pd.read_csv(os.path.join("..", "data", "southern_ocean_surface.csv"))
-X_predict.set_index(['lat', 'lon'], inplace=True)
+X_predict = pd.read_csv(os.path.join("..", "data", "bats_prediction.csv"))
+X_predict.set_index(['depth', 'time'], inplace=True)
+X_predict = X_predict[predictors]
+
+#plotting the prediction data
+vars_list=[('temperature','Temperature','°C'),('PAR', r'PAR ',r'W m$^{-2}$'),('no3','Nitrate','µmol kg$^{-1}$'),('TA','Total Alkalinity','µmol kg$^{-1}$')]
+fig,axs=plt.subplots(4,1,figsize=(7,8),constrained_layout=True)
[…]
