### 2025-09-03 c050150
YAML for BATS example
