### 2025-09-03 5600801
Update example to model BATS
