### 2025-09-03 07be7be
update 2-phase example to focus on SO
