### 2025-09-03 3679466
update YAML for SO
