### 2025-09-03 8a8b0b6
add description of SO dataset
