### 2025-09-04 28a9152
temp function for data processing
