+++ docs/examples/process_bats.py
@@ -1,0 +1,181 @@
+import xarray as xr
+import pandas as pd
+import numpy as np
+import matplotlib.pyplot as plt
+import os
+import xarray as xr
+import pandas as pd
+import numpy as np
+
+# ---- Constants for BATS ----
+BATS_LAT = 31.667
+BATS_LON = -64.167   # west is negative in [-180, 180]
+
+def _normalize_lon_to_180(lon):
+    return ((lon + 180) % 360) - 180
+
+def _find_depth_name(ds_or_df):
+    # Common depth names to accept
+    candidates = ["depth", "Depth", "DEPTH", "z", "lev"]
+    if isinstance(ds_or_df, xr.Dataset):
+        for c in candidates:
+            if c in ds_or_df.dims or c in ds_or_df.coords:
+                return c
+    else:  # DataFrame
+        for c in candidates:
+            if c in ds_or_df.columns:
+                return c
+    raise ValueError("Could not find a depth column/dimension among: " + ", ".join(candidates))
+
+def preprocess_bats(
+    data,
+    lat_col="lat",
+    lon_col="lon",
+    window_deg=0.5,
+    bin_size=25.0,
+):
+    """
+    Subset a box around BATS, BIN DEPTH into fixed layers (25 m by default),
+    and average over lat/lon so the result keeps only (time, depth).
+
+    Depths are returned as negative (surface = 0, deeper = negative).
+
+    Parameters
+    ----------
+    data : xr.Dataset or pd.DataFrame
+        Input data.
+    lat_col, lon_col : str
+        Column names for DataFrame inputs.
+    window_deg : float
+        Half-width of the BATS box in degrees.
+    bin_size : float
+        Depth bin size in meters.
+
+    Returns
+    -------
+    xr.Dataset or pd.DataFrame
+        Averaged over lat/lon with depth binned. Remaining dims/columns are
+        typically ('time','depth').
+    """
+    # -------- xarray case --------
+    if isinstance(data, xr.Dataset):
+        ds = data
+
+        if "lon" not in ds.coords or "lat" not in ds.coords:
+            raise ValueError("xarray.Dataset must have 'lat' and 'lon' coordinates.")
+        depth_name = _find_depth_name(ds)
+
+        # Normalize longitude if needed
+        if float(ds.lon.max()) > 180:
+            ds = ds.assign_coords(lon=_normalize_lon_to_180(ds.lon)).sortby("lon")
+
+        # Subset to BATS box
+        lat_min, lat_max = BATS_LAT - window_deg, BATS_LAT + window_deg
+        lon_min, lon_max = BATS_LON - window_deg, BATS_LON + window_deg
+        ds_box = ds.sel(lat=slice(lat_min, lat_max), lon=slice(lon_min, lon_max))
+
+        # ---- Depth binning ----
+        if depth_name not in ds_box.dims and depth_name not in ds_box.coords:
+            raise ValueError(f"Depth coordinate '{depth_name}' not found after subsetting.")
+
+        depth_vals = ds_box[depth_name].values
+        dmin = float(np.nanmin(depth_vals))
+        dmax = float(np.nanmax(depth_vals))
+        start = bin_size * np.floor(dmin / bin_size)
+        stop = bin_size * np.ceil(dmax / bin_size) + bin_size
+        edges = np.arange(start, stop + 1e-9, bin_size)
+        centers = (edges[:-1] + edges[1:]) / 2.0
+
+        ds_binned = (
+            ds_box
+            .groupby_bins(ds_box[depth_name], bins=edges, labels=centers, right=False)
+            .mean(dim=depth_name, skipna=True, keep_attrs=True)
+        )
+
[…]
