### 2025-09-04 cd4046c
update integrated_totals
