+++ abil/post.py
@@ -744,81 +744,103 @@ class post:
             return total
 
 
-        def integrated_totals(self, targets=None, monthly=False, subset_depth=None, 
-                             export=True, model="ens"):
+        def integrate_total(self, variable='total', monthly=False, subset_depth=None):
             """
-            Estimates global integrated values for all targets.
-    
-            Considers latitude and depth bin size.
-    
+            Estimates global integrated values for a single target. Returns the depth integrated annual total.
+            
+            Works with any subset/ordering of ('lat','lon','depth','time').
+
             Parameters
             ----------
-            targets : an np.array of str, optional
-                An np.array of target variable names to include in the merge. If None, the default 
-                targets from `self.targets` are used (default is None).
+            variable : str
+                The field to be integrated. Default is 'total' from PIC or POC Abil output.
 
             monthly : bool
                 Whether or not to calculate a monthly average value instead of an annual total. Default is False.
- 
+
             subset_depth : float
                 Depth in meters from surface to which integral should be calculated. Default is None. Ex. 100 for top 100m integral.
 
-            export : bool
-                Whether of not to export integrated totals as .csv. Default is True.
-
-            model : str
-                The model version to be integrated. Default is "ens". Other options include {"rf", "xgb", "knn"}.
-    
+            Examples
+            --------
+            >>> m = post(model_config)
+            >>> int = m.Integration(m, resolution_lat=1.0, resolution_lon=1.0, depth_w=5, vol_conversion=1, magnitude_conversion=1e-21, molar_mass=12.01, rate=True)
+            >>> result = integration.integrate_total(variable='Calcification')
+            >>> print("Final integrated total:", result.values)
             """
+            print("Initiate integrated_total")
             ds = self.parent.d.to_xarray()
-            if targets is None:
-                targets = self.targets
-            if "total" in ds:
-                targets = np.append(targets, 'total')
-            if "mean" in ds:
-                targets = np.append(targets, 'mean')
-            if "stdev" in ds:
-                targets = np.append(targets, 'stdev')
-            if "prctile_2.5" in ds:
-                targets = np.append(targets, 'prctile_2.5')
-            if "prctile_97.5" in ds:
-                targets = np.append(targets, 'prctile_97.5')
-            totals = []
+            vol_conversion = self.vol_conversion
+            magnitude_conversion = self.magnitude_conversion
+            molar_mass = self.molar_mass
+            rate = self.rate
 
-            for target in targets:
-                try:
-                    print(f"Processing target: {target}")
-                    total = self.integrate_total(variable=target, monthly=monthly, subset_depth=subset_depth)
-                    total_df = pd.DataFrame({'total': [total.values], 'variable': target})
-                    totals.append(total_df)
-                except Exception as e:
-                    print(f"Some targets do not have predictions! Missing: {target}")
-                    print(f"Error: {e}")
-            totals = pd.concat(totals)
+            # Average number of days for each month (accounting for leap years)
+            days_per_month_full = np.array([31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31])
+            
+            if variable not in ds:
+                raise KeyError(f"Variable '{variable}' not found in dataset.")
 
-            if export:
-                depth_str = f"_depth_{subset_depth}m" if subset_depth else ""
-                month_str = "_monthly_int" if monthly else ""
-                try: #make new dir if needed
-                    os.makedirs(os.path.join(self.parent.root, self.parent.model_config['path_out'], self.parent.model_config['run_name'], "posts/integrated_totals"))
-                except:
-                    None
+            # Optional depth subset if depth exists
+            if subset_depth is not None and ('depth' in ds.dims or 'depth' in ds.coords):
[…]
