+++ abil/post.py
@@ -669,81 +669,7 @@ class post:
             volume.name = 'volume'
             ds['volume'] = volume
             self.parent.d = ds.to_dataframe()
-        
-        def integrate_total(self, variable='total', monthly=False, subset_depth=None):
-            """
-            Estimates global integrated values for a single target. Returns the depth integrated annual total.
             
-            Parameters
-            ----------
-            variable : str
-                The field to be integrated. Default is 'total' from PIC or POC Abil output.
-
-            monthly : bool
-                Whether or not to calculate a monthly average value instead of an annual total. Default is False.
- 
-            subset_depth : float
-                Depth in meters from surface to which integral should be calculated. Default is None. Ex. 100 for top 100m integral.
-
-            Examples
-            --------
-            >>> m = post(model_config)
-            >>> int = m.Integration(m, resolution_lat=1.0, resolution_lon=1.0, depth_w=5, vol_conversion=1, magnitude_conversion=1e-21, molar_mass=12.01, rate=True)
-            >>> result = integration.integrate_total(variable='Calcification')
-            >>> print("Final integrated total:", result.values)
-            """
-            print("Initiate integrated_total")
-            ds = self.parent.d.to_xarray()
-            vol_conversion = self.vol_conversion
-            magnitude_conversion = self.magnitude_conversion
-            molar_mass = self.molar_mass
-            rate = self.rate
-
-            # Average number of days for each month (accounting for leap years)
-            days_per_month_full = np.array([31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31])
-            
-            # Get the available time points (months) from the dataset
-            available_time = ds['time'].values  # Assuming 'time' is an array of 1, 2, 3, and 12
-
-            # Subset the days_per_month array to only include the available months
-            days_per_month = days_per_month_full[available_time - 1]
-
-            if subset_depth:
-                ds = ds.sel(depth=slice(0, subset_depth))
-
-            if rate:
-                if monthly:
-                    # Calculate monthly total (separately for each month)
-                    total = []
-                    for i,month in enumerate(available_time):
-                        monthly_total = (ds[variable].isel(time=i) * ds['volume'].isel(time=i) * days_per_month[i]).sum(dim=['lat', 'lon', 'depth'])
-                        monthly_total = (monthly_total * molar_mass) * vol_conversion * magnitude_conversion
-                        total.append(monthly_total)
-                    total = xr.concat(total, dim="month")
-                    print(f"All monthly totals: {total.values}")
-                else:
-                    # Calculate annual total
-                    total = (ds[variable] * ds['volume'] * days_per_month.mean()).sum(dim=['lat', 'lon', 'depth', 'time'])
-                    total = (total * molar_mass) * vol_conversion * magnitude_conversion
-                    print("Final integrated total:", total.values)
-            else:
-                if monthly:
-                    # Calculate monthly total (separately for each month)
-                    total = []
-                    for i,month in enumerate(available_time):
-                        monthly_total = (ds[variable].isel(time=i) * ds['volume']).isel(time=i).sum(dim=['lat', 'lon', 'depth'])
-                        monthly_total = (monthly_total * molar_mass) * vol_conversion * magnitude_conversion
-                        total.append(monthly_total)
-                    total = xr.concat(total, dim="month")
-                    print(f"All monthly totals: {total.values}")
-                else:
-                    # Calculate annual total
-                    total = (ds[variable] * ds['volume']).sum(dim=['lat', 'lon', 'depth', 'time'])
-                    total = (total * molar_mass) * vol_conversion * magnitude_conversion
[…]
