### 2025-09-04 3cbacf6
update merge_obs
