+++ abil/post.py
@@ -1029,74 +1029,102 @@ class post:
 
         print("exported d to: " + self.path_out + file_name + "_" + self.statistic + self.datatype + ".csv")
 
-    def merge_obs(self, file_name, targets=None):
+    def merge_obs(self, file_name, targets=None, index_cols=None):
         """
         Merge model output with observational data and calculate residuals.
 
-        This function integrates model predictions with observational data based on 
-        spatial and temporal indices, calculates residuals, and exports the merged dataset.
+        Dimension-agnostic: joins on whatever shared coordinate columns exist between
+        model output and observations (e.g., lat/lon, lat/lon/time, lat/lon/depth/time,
+        or any other combo).
 
         Parameters
         ----------
         file_name : str
-            The base name of the output file to save the merged dataset.
-        targets : an np.array of str, optional
-            An np.array of target variable names to include in the merge. If None, the default 
-            targets from `self.targets` are used (default is None).
+            Base name of the output file (without extension).
+        targets : array-like of str, optional
+            Target variable names to include. Defaults to self.targets.
+        index_cols : array-like of str, optional
+            Explicit columns to join on. If omitted, they are inferred as the intersection
+            of columns (excluding target/value columns).
 
         Notes
         -----
-        - The function matches the observational data with model predictions based on the 
-        indices `['lat', 'lon', 'depth', 'time']`.
-        - Residuals are calculated as `observed - predicted` for each target variable.
-        - Columns included in the output are the original targets, their modeled values 
-        (suffixed with `_mod`), and their residuals (suffixed with `_resid`).
-        - The merged dataset is saved as a CSV file with a suffix `_PI` followed by the 
-        `pi` value, appended to the output file name.
-        - Observational data is loaded from the path defined in `self.model_config['training']`.
-
-        Raises
-        ------
-        FileNotFoundError
-            If the observational dataset file cannot be found at the specified location.
+        - Residuals are computed as observed - predicted for each target available on both sides.
+        - Output columns include the targets, *_mod, and *_resid. Coordinate columns are used
+        as an index in the CSV (like your original).
         """
-        # Select and rename the target columns for d
+        # ----- Prepare inputs -----
         if targets is None:
-            targets = self.targets
-        d = self.d[targets]
+            targets = list(self.targets)
+        else:
+            targets = list(targets)
 
-        mod_columns = {target: target + '_mod' for target in targets}
-        d = d.rename(mod_columns, axis=1)
-        d.reset_index(inplace=True)
-        d.set_index(['lat', 'lon', 'depth', 'time'], inplace=True)        
+        # Model dataframe: bring coordinates out of the index so we can infer join keys
+        d_mod = self.d.copy()
+        if d_mod.index.name is not None or isinstance(d_mod.index, (pd.MultiIndex, pd.DatetimeIndex)):
+            d_mod = d_mod.reset_index()
 
-        # Read the training targets from the training.csv file defined in model_config
+        # Keep everything for key inference, but rename target cols to *_mod
+        available_model_targets = [t for t in targets if t in d_mod.columns]
+        mod_map = {t: f"{t}_mod" for t in available_model_targets}
+        d_mod = d_mod.rename(columns=mod_map)
+
+        # ---- Read observations ----
         try:
-            df2_path = os.path.join(self.root, self.model_config['training'])
-
+            df2_path = os.path.join(self.root, self.model_config["training"])
             df2 = pd.read_csv(df2_path)
[…]
