### 2025-09-04 2c2ade9
tidy merge_obs
