### 2025-09-04 4e6c81f
add integrated totals to example
