### 2025-09-04 21bd863
expand rst to include integrated totals
