### 2025-09-04 9ea553d
change prediction plots to abundance
