+++ docs/examples/2-phase.py
@@ -25,6 +25,7 @@ os.chdir(os.path.join(".", "docs", "examples"))
 #load configuration yaml:
 with open('2-phase.yml', 'r') as f:
     model_config = load(f, Loader=Loader)
+
 #load example training data:
 d = pd.read_csv(os.path.join("..", "data", "so_training.csv"))
 #define target:
@@ -35,26 +36,25 @@ predictors = model_config['predictors']
 y = d[target]
 X_train = d[predictors]
 
-#train your model:
-m = tune(X_train, y, model_config)
-m.train(model="rf")
-m.train(model="xgb")
-m.train(model="knn")
+# #train your model:
+# m = tune(X_train, y, model_config)
+# m.train(model="rf")
+# m.train(model="xgb")
+# m.train(model="knn")
 
 #load prediction data:
 X_predict = pd.read_csv(os.path.join("..", "data", "so_prediction.csv"))
 X_predict.set_index(['lat', 'lon'], inplace=True)
 X_predict = X_predict[predictors]
 
-#predict your model:
-m = predict(X_train, y, X_predict, model_config)
-m.make_prediction()
+# #predict your model:
+# m = predict(X_train, y, X_predict, model_config)
+# m.make_prediction()
 
 # Posts
 targets = np.array([target])
 def do_post(statistic):
-    m = post(X_train, y, X_predict, model_config, statistic, datatype="poc")
-    m.estimate_carbon("pg poc")
+    m = post(X_train, y, X_predict, model_config, statistic, datatype="abundance")
     m.export_ds("my_first_2-phase_model")
 
 do_post(statistic="mean")
@@ -62,9 +62,9 @@ do_post(statistic="ci95_UL")
 do_post(statistic="ci95_LL")
 
 # Load the predictions
-ds = xr.open_dataset("./ModelOutput/2-phase/posts/my_first_2-phase_model_mean_poc.nc")
-ds_UL = xr.open_dataset("./ModelOutput/2-phase/posts/my_first_2-phase_model_ci95_UL_poc.nc")
-ds_LL = xr.open_dataset("./ModelOutput/2-phase/posts/my_first_2-phase_model_ci95_LL_poc.nc")
+ds = xr.open_dataset("./ModelOutput/2-phase/posts/my_first_2-phase_model_mean_abundance.nc")
+ds_UL = xr.open_dataset("./ModelOutput/2-phase/posts/my_first_2-phase_model_ci95_UL_abundance.nc")
+ds_LL = xr.open_dataset("./ModelOutput/2-phase/posts/my_first_2-phase_model_ci95_LL_abundance.nc")
 
 # Create the figure
 def plot_panel(ax, data, var, title, label, cmap='viridis', cbar_label='abundance (cells L$^{-1}$)'):
@@ -72,11 +72,11 @@ def plot_panel(ax, data, var, title, label, cmap='viridis', cbar_label='abundanc
     ax.add_feature(cfeature.LAND, color='gray'); ax.add_feature(cfeature.COASTLINE, linewidth=0.5)
     ax.set_title(f'$\mathbf{{{label}}}$  {title}', loc='left', pad=10, y=1.05, fontsize=10)
 
-    if isinstance(data, pd.DataFrame):  # raw points -> hexbin (no xarray conversion)
+    if isinstance(data, pd.DataFrame):  # raw points
         lon, lat, vals = np.asarray(data['lon']), np.asarray(data['lat']), np.asarray(data[var])
         vmin, vmax = np.nanpercentile(vals, [2,98]); x, y = ax.projection.transform_points(ccrs.PlateCarree(), lon, lat)[:,0:2].T
[…]
