### 2025-09-04 e150e16
fix rst with new line breaks
