+++ docs/source/user-guide/two-phase-ensemble.rst
@@ -1,10 +1,10 @@
 2-phase Ensemble 
 ****************
 
-In this example we will use Abil to predict the biomass of a highly abundant calcifying nanoplankton (`Gephyrocapsa huxleyi`).
-We will focus on the Southern Ocean, a region with high `G. huxleyi`` stocks and of significant importance to the ocean carbon cycle.
-This region has an interesting macroscale distribution of calcifying nanoplankton biomass, with peak concentrations in the so called "great calcite belt" 
-between 40-60°S but a notable absence below 60° due to competition with silicifying nanoplankton.
+In this example we will use Abil to predict the biomass of a highly abundant calcifying nanoplankton (*Gephyrocapsa huxleyi*).
+We will focus on the Southern Ocean, a region with high *G. huxleyi* stocks and of significant importance to the ocean carbon cycle.
+This region has an interesting macroscale distribution of *G. huxleyi* biomass, with peak concentrations in the so called "great calcite belt" 
+between ~40-60°S but a notable absence below ~60°S due to competition with silicifying nanoplankton.
 
 YAML example
 ~~~~~~~~~~~~
@@ -15,7 +15,7 @@ For a detailed explanation of each parameter see :ref:`yaml_config`.
 An example of YAML file of a 2-phase model is provided below.
 Note that compared to a 1-phase regressor model, the hyper-parameters for the classifier also need to be specified.
 
-.. literalinclude:: ../../../tests/2-phase.yml
+.. literalinclude:: ../../examples/2-phase.yml
    :language: yaml
 
 Running the model
@@ -35,28 +35,27 @@ For instructions on how to install these packages, see `requirements.txt <../../
 and the Abil :ref:`getting-started`.
 
 .. literalinclude:: ../../examples/2-phase.py
-   :lines: 4-10
+   :lines: 4-2``
    :language: python
 
 Loading the configuration YAML
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 After loading the required packages we need to define our file paths.
-Note that this is operating system specific, as Unix and Mac use '/' while for Windows '\' is used.
 
 .. literalinclude:: ../../examples/2-phase.py
-   :lines: 16-17
+   :lines: 23
    :language: python
 
 
-Creating example data
-^^^^^^^^^^^^^^^^^^^^^
+Loading example data
+^^^^^^^^^^^^^^^^^^^^
 
-Next we create some example data. When applying the pipeline to your own data, note that the data
+Next we load our Southern Ocean example data. When applying the pipeline to your own data, note that the data
 needs to be in a `Pandas DataFrame format <https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.html>`_.
 
 .. literalinclude:: ../../examples/2-phase.py
-   :lines: 20-23
+   :lines: 29-37
    :language: python
 
 Training the model
@@ -66,7 +65,7 @@ Next we train our model. Note that depending on the number of hyper-parameters s
 YAML file this can be computationally very expensive and it recommended to do this on a HPC system. 
 
 .. literalinclude:: ../../examples/2-phase.py
-   :lines: 26-29
+   :lines: 39-43
    :language: python
 
 Making predictions
@@ -75,7 +74,7 @@ Making predictions
 After training our model we can make predictions on a new dataset (X_predict):
 
 .. literalinclude:: ../../examples/2-phase.py
-   :lines: 32-33
+   :lines: 45-52
    :language: python
 
 Post-processing
@@ -84,7 +83,7 @@ Post-processing
 Finally, we conduct the post-processing.
 
 .. literalinclude:: ../../examples/2-phase.py
-   :lines: 36-58
+   :lines: 55-62
[…]
