+++ docs/source/user-guide/one-phase-ensemble.rst
@@ -2,8 +2,7 @@ Coccolithophore abundance in the Bermuda Atlantic Time Series
 *************************************************************
 
 In this example we will use Abil to predict the biomass of the calcifying nanoplankton ('Coccolithophores').
-
-This species is highly abundant in the global oceans, and plays an important role in the carbon cycle.
+We will focus on the species: *Gephyrocapsa huxleyi* which is highly abundant in the global oceans, and plays an important role in the carbon cycle.
 We fill focus this example on the Bermuda Atlantic Time Series (BATS), a location with great temporal coverage of *G. huxleyi*.
 In this location, wind-driven mixing in the winter drives nutrient upwelling which results in high concentrations of *G. huxleyi*.
 In the summer, as nutrient influx decreases due to lower mixing and subsequently the abundance of *G. huxleyi* decreases. 
