### 2025-09-04 5035d11
implement by manual specification of index_cols instead
