+++ docs/source/user-guide/one-phase-ensemble.rst
@@ -2,11 +2,14 @@ Coccolithophore abundance in the Bermuda Atlantic Time Series
 *************************************************************
 
 In this example we will use Abil to predict the biomass of the calcifying nanoplankton ('Coccolithophores').
-We will focus on the species: *Gephyrocapsa huxleyi* which is highly abundant in the global oceans, and plays an important role in the carbon cycle.
+We will focus on the species *Gephyrocapsa huxleyi* which is highly abundant in the global oceans, and plays an important role in the carbon cycle.
 We fill focus this example on the Bermuda Atlantic Time Series (BATS), a location with great temporal coverage of *G. huxleyi*.
 In this location, wind-driven mixing in the winter drives nutrient upwelling which results in high concentrations of *G. huxleyi*.
-In the summer, as nutrient influx decreases due to lower mixing and subsequently the abundance of *G. huxleyi* decreases. 
+In the summer, as nutrient influx decreases due to lower mixing the abundance of *G. huxleyi* subsequently decreases :footcite:p:`haidar2001coccolithophore`. 
 An interesting temporal dynamic we will try to capture with our machine learning ensemble. 
+
+As *Gephyrocapsa huxleyi* is present in low concentrations most of the year, we will use a 1-phase regression model.
+For an example of the 2-phase model see :doc:`Southern Ocean distribution of Gephyrocapsa huxleyi <two-phase-ensemble>`. 
 
 YAML example
 ~~~~~~~~~~~~
@@ -37,7 +40,7 @@ For instructions on how to install these packages, see `requirements.txt <../../
 and the Abil :ref:`getting-started`.
 
 .. literalinclude:: ../../examples/regressor.py
-   :lines: 4-17
+   :lines: 4-18
    :language: python
 
 Loading the configuration YAML
@@ -46,28 +49,29 @@ Loading the configuration YAML
 After loading the required packages we need to define our file paths.
 
 .. literalinclude:: ../../examples/regressor.py
-   :lines: 19
+   :lines: 20
    :language: python
 
 Then we can load the YAML:
 
 .. literalinclude:: ../../examples/regressor.py
-   :lines: 21-23
+   :lines: 22-24
    :language: python
 
 Loading example data
 ^^^^^^^^^^^^^^^^^^^^^
 
-Next we load some example data, here we utilize abundance data from the CASCADE database (10.5281/zenodo.12797197).
+Next we load some example data, here we utilize abundance data from the CASCADE database :footcite:p:`deVries2024cascade`.
 The CASCADE database provides observations gridded to 1 degree x 1 degree x 5 meters x 1 month. 
 For the example we have subset the data for Bermuda and have averaged the observations with latitude and longitude. 
 In addition to our predictors (`y_train`) we also need environmental data which match our predictions ('X_train'). 
-This data was obtained from monthly climatologies from data sources such as the World Ocean Atlas, and NNGv2.
+This data was obtained from monthly climatologies from data sources such as the World Ocean Atlas :footcite:p:`Reagan2024`, NNGv2  :footcite:p:`Broullon2019, Broullon2020` and Castant et al., 2024  :footcite:p:`Castant2024`.
+
 When applying the pipeline to your own data, note that the data
 needs to be in a `Pandas DataFrame format <https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.html>`_.
 
 .. literalinclude:: ../../examples/regressor.py
-   :lines: 24-32
+   :lines: 26-34
[…]

[…]
