### 2025-09-05 8635cf0
run examples as part of doc CI
