+++ docs/source/user-guide/one-phase-ensemble.rst
@@ -19,7 +19,7 @@ For a detailed explanation of each parameter see :ref:`yaml_config`.
 
 An example of YAML file of a 1-phase model is provided below.
 
-.. literalinclude:: ../../examples/regressor.yml
+.. literalinclude:: ../../../examples/regressor.yml
    :language: yaml
 
 
@@ -39,7 +39,7 @@ Before running the Python script we need to import all relevant Python packages.
 For instructions on how to install these packages, see `requirements.txt <../../../../../requirements.txt>`_
 and the Abil :ref:`getting-started`.
 
-.. literalinclude:: ../../examples/regressor.py
+.. literalinclude:: ../../../examples/regressor.py
    :lines: 4-18
    :language: python
 
@@ -48,13 +48,13 @@ Loading the configuration YAML
 
 After loading the required packages we need to define our file paths.
 
-.. literalinclude:: ../../examples/regressor.py
+.. literalinclude:: ../../../examples/regressor.py
    :lines: 20
    :language: python
 
 Then we can load the YAML:
 
-.. literalinclude:: ../../examples/regressor.py
+.. literalinclude:: ../../../examples/regressor.py
    :lines: 22-24
    :language: python
 
@@ -70,7 +70,7 @@ This data was obtained from monthly climatologies from data sources such as the 
 When applying the pipeline to your own data, note that the data
 needs to be in a `Pandas DataFrame format <https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.html>`_.
 
-.. literalinclude:: ../../examples/regressor.py
+.. literalinclude:: ../../../examples/regressor.py
    :lines: 26-34
    :language: python
 
@@ -79,11 +79,11 @@ Plotting example data
 
 We can have a look at the example data. It already has pretty good data coverage.
 
-.. literalinclude:: ../../examples/regressor.py
+.. literalinclude:: ../../../examples/regressor.py
    :lines: 36-54
    :language: python
 
-.. figure:: ../../examples/BATS_observations.png
+.. figure:: ../../../examples/BATS_observations.png
 
 Training the model
 ^^^^^^^^^^^^^^^^^^
@@ -91,7 +91,7 @@ Training the model
 Next we train our model. Note that depending on the number of hyper-parameters specified in the
 YAML file this can be computationally very expensive and it recommended to do this on a HPC system. 
 
-.. literalinclude:: ../../examples/regressor.py
+.. literalinclude:: ../../../examples/regressor.py
    :lines: 56-60
    :language: python
 
@@ -102,23 +102,23 @@ After training our model we can make predictions on the BATS environmental datas
 
 First we need to load our environmental data to make the predictions on (X_predict):
 
-.. literalinclude:: ../../examples/regressor.py
+.. literalinclude:: ../../../examples/regressor.py
    :lines: 62-65
    :language: python
 
 We can also quickly plot the environmental data. Note the seasonality in key parameters such as PAR and temperature.
 
 
-.. literalinclude:: ../../examples/regressor.py
+.. literalinclude:: ../../../examples/regressor.py
    :lines: 67-77
    :language: python
 
-.. figure:: ../../examples/BATS_environment.png
+.. figure:: ../../../examples/BATS_environment.png
 
 
 Then we can make our predictions:
 
-.. literalinclude:: ../../examples/regressor.py
+.. literalinclude:: ../../../examples/regressor.py
    :lines: 79-81
    :language: python
 
@@ -127,7 +127,7 @@ Post-processing
 
 Finally, we conduct the post-processing.
 
-.. literalinclude:: ../../examples/regressor.py
[…]
