### 2025-09-05 7c791ed
temporarily build docs on pull_request
