+++ .github/workflows/deploy.yml
@@ -4,7 +4,9 @@ on:
   push:
     branches:
       - main  
-
+  pull_request:
+    branches:
+      - main
 jobs:
   build:
     runs-on: ubuntu-latest
