### 2025-09-05 77ad46d
add missing doc deps
