+++ .github/workflows/deploy.yml
@@ -27,6 +27,8 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt  # Ensure your requirements.txt includes Sphinx
+          pip install matplotlib
+          pip install seaborn
           pip install sphinx
           pip install pydata-sphinx-theme
           pip install sphinx-design
