+++ .github/workflows/deploy.yml
@@ -4,9 +4,7 @@ on:
   push:
     branches:
       - main  
-  pull_request:
-    branches:
-      - main
+
 jobs:
   build:
     runs-on: ubuntu-latest
@@ -53,4 +51,4 @@ jobs:
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: ./docs/_build/html  # Adjust according to your Sphinx setup
-#          publish_branch: gh-pages  # The branch GitHub Pages usesnanophyto/abil-docs  # The public repository your-username/public-repo  # Replace with your public repository
+          publish_branch: gh-pages  # The branch GitHub Pages usesnanophyto/abil-docs  # The public repository your-username/public-repo  # Replace with your public repository
