### 2025-09-05 8ca4bd7
add missing definitions of seed
