### 2025-09-05 a9cbbcc
add shuffle option to avoid unused seed warning
