### 2025-09-05 35bdf2b
add initial assertion test
