### 2025-09-05 b9f449b
add random_state to prediction quantiles
