+++ abil/predict.py
@@ -79,7 +79,7 @@ def load_model_and_scores(path_out, ensemble_config, n, target):
     return(m, scores)
 
 
-def export_prediction(ensemble_config, m, target, target_no_space, X_predict, X_train, y_train, cv, model_out, n_threads=8):
+def export_prediction(ensemble_config, m, target, target_no_space, X_predict, X_train, y_train, cv, model_out, n_threads=8, random_state=None):
     """
     Exports model predictions to a NetCDF file.
 
@@ -104,7 +104,7 @@ def export_prediction(ensemble_config, m, target, target_no_space, X_predict, X_
     if (ensemble_config["classifier"] ==False) and (ensemble_config["regressor"] == True):
         with parallel_backend("loky", n_jobs=n_threads):
             d = pp.estimate_prediction_quantiles(
-                m, X_predict=X_predict, X_train=X_train, y_train=y_train, cv=cv
+                m, X_predict=X_predict, X_train=X_train, y_train=y_train, cv=cv, random_state=random_state
             )["predict_stats"]
             d['mean'] = m.predict(X_predict)
         
@@ -335,7 +335,7 @@ class predict:
             model_name = self.ensemble_config["m" + str(1)]
             model_out = os.path.join(self.path_out, "predictions", model_name)
             export_prediction(self.ensemble_config, m, self.target, self.target_no_space, self.X_predict, self.X_train, self.y, self.cv, 
-                              model_out, n_threads=self.n_jobs)
+                              model_out, n_threads=self.n_jobs, random_state=self.seed)
 
         elif number_of_models >=2:
                     
@@ -351,11 +351,11 @@ class predict:
 
                 if (self.ensemble_config["classifier"] ==False) and (self.ensemble_config["regressor"] == True):
                     export_prediction(self.ensemble_config, m, self.target, self.target_no_space, self.X_predict, self.X_train, self.y, self.cv, 
-                                    model_out, n_threads=self.n_jobs)
+                                    model_out, n_threads=self.n_jobs, random_state=self.seed)
                 if (self.ensemble_config["classifier"] ==True) and (self.ensemble_config["regressor"] == True):
 
                     export_prediction(self.ensemble_config, m, self.target, self.target_no_space, self.X_predict, self.X_train, self.y, self.cv, 
-                                     model_out, n_threads=self.n_jobs)
+                                     model_out, n_threads=self.n_jobs, random_state=self.seed)
 
                 models.append((model_name, m))
                 mae_values.append(mae)
@@ -365,7 +365,7 @@ class predict:
                 m = VotingRegressor(estimators=models, weights=w).fit(self.X_train, self.y)   
                 model_out = os.path.join(self.path_out, "predictions", "ens")
                 export_prediction(self.ensemble_config, m, self.target, self.target_no_space, self.X_predict, self.X_train, self.y, self.cv, 
-                                model_out, n_threads=self.n_jobs)
+                                model_out, n_threads=self.n_jobs, random_state=self.seed)
                 
                 #export model object:
                 base_output_path = os.path.join(self.path_out, "model", "ens")
@@ -412,7 +412,7 @@ class predict:
                 m.fit(self.X_train, y)
                 model_out = os.path.join(self.path_out, "predictions", "ens")
                 export_prediction(self.ensemble_config, m, self.target, self.target_no_space, self.X_predict, self.X_train, self.y, self.cv, 
-                                model_out, n_threads=self.n_jobs)
+                                model_out, n_threads=self.n_jobs, random_state=self.seed)
                 base_output_path = os.path.join(self.path_out, "model", "ens")
 
                 try: #make new dir if needed
