### 2025-09-08 a3d32a2
add check for mean
