### 2025-09-08 ebe73f3
assert total for regressor
