+++ requirements.txt
@@ -1,10 +1,10 @@
-netcdf4
-dask
+netcdf4==1.6.5
+dask==2024.5.0
 numpy==1.26.4
-pandas
-pyyaml
-xarray
-xgboost
-scikit-learn<1.6
-scipy
+pandas==2.2.2
+pyyaml==6.0.1
+xarray==2023.6.0
+xgboost==1.7.6
+scikit-learn==1.3.0
+scipy==1.13.1
 scikit-bio
