### 2025-09-08 90732a8
update requirements with versions that work with scikit-bio
