+++ requirements.txt
@@ -1,10 +1,10 @@
-dask>=2022.2.0
-joblib>=1.3.2
-netcdf4>=1.6.5
-numpy>=1.21.6
-pandas>=1.1.5
-pyyaml>=6.0.1
-scikit-learn>=1.0.2,<1.6
-scipy>=1.7.3
-xarray>=0.20.2
-xgboost>=1.6.2
+netcdf4==1.6.5
+dask==2024.5.0
+numpy==1.26.4
+pandas==2.2.2
+pyyaml==6.0.1
+xarray==2023.6.0
+xgboost==1.7.6
+scikit-learn==1.3.0
+scipy==1.13.1
+scikit-bio==0.6.2
