### 2025-09-11 f62096d
move to fstrings consistently
