+++ abil/post.py
@@ -11,6 +11,8 @@ from skbio.diversity.alpha import shannon
 
 from .analyze import area_of_applicability
 
+logger = logging.getLogger("abil")
+
 class AbilPostProcessor:
     """
     Post processing of SDM
@@ -77,7 +79,7 @@ class AbilPostProcessor:
 
             This function uses `xarray.open_mfdataset` to load all NetCDF files in the given directory 
             (matching the pattern "*.nc") and combines them into one xarray.Dataset. The function 
-            logging.info(s status messages indicating the start and completion of the merging process.
+            logs status messages indicating the start and completion of the merging process.
 
             Parameters
             ----------
@@ -92,8 +94,7 @@ class AbilPostProcessor:
                 The merged dataset containing the combined data from all the NetCDF files in the directory.
                 The variable names in the merged dataset are derived from the 'target' values in each file.
             """
-            logging.info("merging...")
-            logging.info(path_in)
+            logger.info(f"merging... {path_in}")
 
             datasets = []
             
@@ -108,12 +109,12 @@ class AbilPostProcessor:
                         ds_subset = ds[[statistic]].rename({statistic: target_name})
                         datasets.append(ds_subset)
                     else:
-                        logging.info(f"Statistic '{statistic}' not found in {file}")
+                        logger.info(f"Statistic '{statistic}' not found in {file}")
 
             # Merge datasets by variables, keeping same coordinates
             merged_ds = xr.merge(datasets, compat='override')  # 'override' skips conflicts
 
-            logging.info("finished merging NetCDF files")
+            logger.info("finished merging NetCDF files")
             return merged_ds
 
         self.path_out = os.path.join(model_config['root'], model_config['path_out'], model_config['run_name'], "posts/")
@@ -163,7 +164,7 @@ class AbilPostProcessor:
         ------
         Exception
             If an error occurs during the directory creation or file writing process, an exception
-            is caught and an error message is logging.info(ed.
+            is caught and an error message is logger.info(ed.
 
         Notes
         -----
@@ -177,9 +178,9 @@ class AbilPostProcessor:
             with open(yml_file_path, 'w') as yml_file:
                 dump(self.model_config, yml_file, Dumper=Dumper, default_flow_style=False)
             
-            logging.info(f"Model configuration exported to: {yml_file_path}")
+            logger.info(f"Model configuration exported to: {yml_file_path}")
         except Exception as e:
-            logging.info(f"Error exporting model_config to YAML: {e}")   
+            logger.info(f"Error exporting model_config to YAML: {e}")   
 
     def merge_performance(self):
         """
@@ -192,8 +193,7 @@ class AbilPostProcessor:
         """    
 
         models = [value for key, value in self.model_config['ensemble_config'].items() if key.startswith("m")]
-        logging.info("models included in merge performance!")
-        logging.info(models)
+        logger.info(f"models included in merge performance! {models}")
         models.append("ens")
         for model in models:
             self.merge_performance_single_model(model)
@@ -261,7 +261,7 @@ class AbilPostProcessor:
         os.makedirs(os.path.join(self.root, self.model_config['path_out'], self.model_config['run_name'], "posts/performance"), exist_ok=True)
         all_performance.to_csv(os.path.join(self.root, self.model_config['path_out'], self.model_config['run_name'], "posts/performance", model) + "_performance.csv", index=False)
 
-        logging.info("finished merging performance")
+        logger.info("finished merging performance")
 
     def merge_parameters(self):
         """
[…]
