+++ abil/unified_tree_or_bag.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pandas as pd
 import logging
+logger = logging.getLogger("abil")
 
 from sklearn.metrics import mean_absolute_error, balanced_accuracy_score
 
@@ -510,11 +511,11 @@ if __name__ == "__main__":
         vzir_results = estimate_prediction_quantiles(
             zir_of_vmodels, X_predict=X_predict, X_train=X_train, y_train=y_train, cv=cv
         )
-    logging.info("\n=== Training Summary Stats ===\n", results["train_stats"].head())
-    logging.info("\n=== Prediction Summary Stats ===\n", results["predict_stats"].head())
+    logging.info(f"\n=== Training Summary Stats ===\n{results['train_stats'].head()}")
+    logging.info(f"\n=== Prediction Summary Stats ===\n{results['predict_stats'].head()}")
 
-    logging.info("\n=== Training Summary Stats ===\n", vzir_results["classifier_train_stats"].head())
-    logging.info("\n=== Prediction Summary Stats ===\n", vzir_results["classifier_predict_stats"].head())
+    logging.info(f"\n=== Training Summary Stats ===\n{vzir_results['classifier_train_stats'].head()}")
+    logging.info(f"\n=== Prediction Summary Stats ===\n{vzir_results['classifier_predict_stats'].head()}")
     
-    logging.info("\n=== Training Summary Stats ===\n", vzir_results["regressor_train_stats"].head())
-    logging.info("\n=== Prediction Summary Stats ===\n", vzir_results["regressor_predict_stats"].head())
+    logging.info(f"\n=== Training Summary Stats ===\n{vzir_results['regressor_train_stats'].head()}")
+    logging.info(f"\n=== Prediction Summary Stats ===\{vzir_results['regressor_predict_stats'].head()}")
