### 2025-09-11 8362652
add explicit refs to examples in manuscript
