### 2025-09-11 910ffca
more consistent bibtex formatting
