+++ docs/library.bib
@@ -1,6 +1,6 @@
 @article{deVries2024cascade,
   title={CASCADE: Dataset of extant coccolithophore size, carbon content and global distribution},
-  author={De Vries, Joost and Poulton, Alex J and Young, Jeremy R and Monteiro, Fanny M and Sheward, Rosie M and Johnson, Roberta and Hagino, Kyoko and Ziveri, Patrizia and Wolf, Levi J},
+  author={De Vries, J. and Poulton, A. J. and Young, J. R. and Monteiro, F. M. and Sheward, R. M. and Johnson, R. and Hagino, K. and Ziveri, P. and Wolf, L. J.},
   journal={Scientific data},
   volume={11},
   number={1},
@@ -10,7 +10,7 @@
 }
 @article{nissen2018factors,
   title={Factors controlling coccolithophore biogeography in the Southern Ocean},
-  author={Nissen, Cara and Vogt, Meike and M{\"u}nnich, Matthias and Gruber, Nicolas and Haumann, F Alexander},
+  author={Nissen, C. and Vogt, M. and M{\"u}nnich, M. and Gruber, N. and Haumann, F. A.},
   journal={Biogeosciences},
   volume={15},
   number={22},
@@ -20,7 +20,7 @@
 }
 @article{haidar2001coccolithophore,
   title={Coccolithophore dynamics off bermuda (n. atlantic)},
-  author={Haidar, Ali T and Thierstein, Hans R},
+  author={Haidar, A. T. and Thierstein, H. R.},
   journal={Deep Sea Research Part II: Topical Studies in Oceanography},
   volume={48},
   number={8-9},
