### 2025-09-11 ca9c97a
add temp bibtex entry
