+++ paper/paper.bib
@@ -240,4 +240,9 @@ version = {latest}
   doi={10.1111/j.1600-0587.2008.05742.x},
   publisher={Wiley Online Library}
 }
-
+@article{devries:2025,
+  title        = {A diverse community constitutes global coccolithophore calcium carbonate stocks},
+  author       = {de Vries, Joost and Monteiro, Fanny M. and Poulton, Alex J. and Wiseman, Nicola A. and Wolf, Levi J.},
+  journal      = {bioRxiv},
+  year         = {2025},
+}
