+++ paper/paper.bib
@@ -242,7 +242,7 @@ version = {latest}
 }
 @article{devries:2025,
   title        = {A diverse community constitutes global coccolithophore calcium carbonate stocks},
-  author       = {de Vries, Joost and Monteiro, Fanny M. and Poulton, Alex J. and Wiseman, Nicola A. and Wolf, Levi J.},
+  author       = {{de Vries}, Joost and Monteiro, Fanny M. and Poulton, Alex J. and Wiseman, Nicola A. and Wolf, Levi J.},
   journal      = {bioRxiv},
   year         = {2025},
 }
