+++ _sources/source/user-guide/one-phase-ensemble.rst.txt
@@ -1,15 +1,5 @@
-Coccolithophore abundance in the Bermuda Atlantic Time Series 
-*************************************************************
-
-In this example we will use Abil to predict the biomass of the calcifying nanoplankton ('Coccolithophores').
-We will focus on the species *Gephyrocapsa huxleyi* which is highly abundant in the global oceans, and plays an important role in the carbon cycle.
-We fill focus this example on the Bermuda Atlantic Time Series (BATS), a location with great temporal coverage of *G. huxleyi*.
-In this location, wind-driven mixing in the winter drives nutrient upwelling which results in high concentrations of *G. huxleyi*.
-In the summer, as nutrient influx decreases due to lower mixing the abundance of *G. huxleyi* subsequently decreases :footcite:p:`haidar2001coccolithophore`. 
-An interesting temporal dynamic we will try to capture with our machine learning ensemble. 
-
-As *Gephyrocapsa huxleyi* is present in low concentrations most of the year, we will use a 1-phase regression model.
-For an example of the 2-phase model see :doc:`Southern Ocean distribution of Gephyrocapsa huxleyi <two-phase-ensemble>`. 
+1-phase Ensemble 
+****************
 
 YAML example
 ~~~~~~~~~~~~
@@ -19,7 +9,7 @@ For a detailed explanation of each parameter see :ref:`yaml_config`.
 
 An example of YAML file of a 1-phase model is provided below.
 
-.. literalinclude:: ../../../examples/regressor.yml
+.. literalinclude:: ../../../tests/regressor.yml
    :language: yaml
 
 
@@ -39,51 +29,30 @@ Before running the Python script we need to import all relevant Python packages.
 For instructions on how to install these packages, see `requirements.txt <../../../../../requirements.txt>`_
 and the Abil :ref:`getting-started`.
 
-.. literalinclude:: ../../../examples/regressor.py
-   :lines: 4-18
+.. literalinclude:: ../../examples/regressor.py
+   :lines: 4-10
    :language: python
 
 Loading the configuration YAML
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 After loading the required packages we need to define our file paths.
+Note that this is operating system specific, as Unix and Mac use '/' while for Windows '\' is used.
 
-.. literalinclude:: ../../../examples/regressor.py
-   :lines: 20
+.. literalinclude:: ../../examples/regressor.py
+   :lines: 16-17
    :language: python
 
-Then we can load the YAML:
 
-.. literalinclude:: ../../../examples/regressor.py
-   :lines: 22-24
-   :language: python
-
-Loading example data
+Creating example data
 ^^^^^^^^^^^^^^^^^^^^^
 
-Next we load some example data, here we utilize abundance data from the CASCADE database :footcite:p:`deVries2024cascade`.
-The CASCADE database provides observations gridded to 1 degree x 1 degree x 5 meters x 1 month. 
-For the example we have subset the data for Bermuda and have averaged the observations with latitude and longitude. 
-In addition to our predictors (`y_train`) we also need environmental data which match our predictions ('X_train'). 
-This data was obtained from monthly climatologies from data sources such as the World Ocean Atlas :footcite:p:`Reagan2024`, NNGv2  :footcite:p:`Broullon2019, Broullon2020` and Castant et al., 2024  :footcite:p:`Castant2024`.
-
-When applying the pipeline to your own data, note that the data
[…]
