+++ _sources/source/user-guide/hpc.rst.txt
@@ -130,7 +130,7 @@ The first part of the bash script declares the variables needed to execute the j
 Here, we include the time limit for the run (time), the number of nodes to use (nodes), the memory allocation (mem),
 the number of cpus per task (cpus-per-task), and the number of targets to be tuned (array).
 
-.. literalinclude:: ../../../examples/tune_KNN.sh
+.. literalinclude:: ../../examples/tune_KNN.sh
     :lines: 1-8
     :language: shell
 .. note::
@@ -142,20 +142,20 @@ Executable commands
 The next part of the bash script includes the commands to be executed.
 First, the array value is used to set a local variable that will be used to specify the target being tuned.
 
-.. literalinclude:: ../../../examples/tune_KNN.sh
+.. literalinclude:: ../../examples/tune_KNN.sh
     :lines: 10
     :language: shell
 
 Next, the Apptainer module is loaded, and set up using the abil.sif container uploaded prior.
 
-.. literalinclude:: ../../../examples/tune_KNN.sh
+.. literalinclude:: ../../examples/tune_KNN.sh
     :lines: 12-16
     :language: shell
 
 Finally, the model Python script is executed using the specified number of cpus, for the target "i", within a specific model (knn in this instance).
 Lastly, the Singularity cache is exported.
 
-.. literalinclude:: ../../../examples/tune_KNN.sh
+.. literalinclude:: ../../examples/tune_KNN.sh
     :lines: 17-19
     :language: shell
 
@@ -165,13 +165,13 @@ Alterations for predict and post
 The set up is the same for each the predict.sh and post.sh scripts, with the only change being the Python executable line.
 predict.sh should say the following:
 
-.. literalinclude:: ../../../examples/predict.sh
+.. literalinclude:: ../../examples/predict.sh
     :lines: 17
     :language: shell
 
 while post.sh should say the following, and does not include the array specification:
 
-.. literalinclude:: ../../../examples/post.sh
+.. literalinclude:: ../../examples/post.sh
     :lines: 14
     :language: shell
 
@@ -222,5 +222,5 @@ Singularity file
 ----------------
 Below is the Singularity.sif file text. This is used to create abil.sif in the steps above.
 
-.. literalinclude:: ../../../examples/Singularity.sif
+.. literalinclude:: ../../examples/Singularity.sif
     :language: Singularity
