+++ _sources/source/user-guide/two-phase-ensemble.rst.txt
@@ -1,13 +1,5 @@
-Southern Ocean distribution of *Gephyrocapsa huxleyi*
-*****************************************************
-
-In this example we will use Abil to predict the biomass of a highly abundant calcifying nanoplankton (*Gephyrocapsa huxleyi*).
-We will focus on the Southern Ocean, a region with high *G. huxleyi* stocks and of significant importance to the ocean carbon cycle.
-This region has an interesting macroscale distribution of *G. huxleyi* biomass, with peak concentrations in the so called "great calcite belt" 
-between ~40-60°S but a notable absence below ~60°S due to competition with silicifying nanoplankton :footcite:p:`nissen2018factors`.
-
-For this example we will use a 2-phase regressor to better constrain absences below ~60°S.
-For an example of the 1-phase model see :doc:`Coccolithophore abundance in the Bermuda Atlantic Time Series  <one-phase-ensemble>`. 
+2-phase Ensemble 
+****************
 
 YAML example
 ~~~~~~~~~~~~
@@ -18,7 +10,7 @@ For a detailed explanation of each parameter see :ref:`yaml_config`.
 An example of YAML file of a 2-phase model is provided below.
 Note that compared to a 1-phase regressor model, the hyper-parameters for the classifier also need to be specified.
 
-.. literalinclude:: ../../../examples/2-phase.yml
+.. literalinclude:: ../../../tests/2-phase.yml
    :language: yaml
 
 Running the model
@@ -37,41 +29,39 @@ Before running the Python script we need to import all relevant Python packages.
 For instructions on how to install these packages, see `requirements.txt <../../../../../requirements.txt>`_
 and the Abil :ref:`getting-started`.
 
-.. literalinclude:: ../../../examples/2-phase.py
-   :lines: 4-21
+.. literalinclude:: ../../examples/2-phase.py
+   :lines: 4-10
    :language: python
 
 Loading the configuration YAML
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 After loading the required packages we need to define our file paths.
+Note that this is operating system specific, as Unix and Mac use '/' while for Windows '\' is used.
 
-.. literalinclude:: ../../../examples/2-phase.py
-   :lines: 23
+.. literalinclude:: ../../examples/2-phase.py
+   :lines: 16-17
    :language: python
 
 
-Loading example data
-^^^^^^^^^^^^^^^^^^^^
+Creating example data
+^^^^^^^^^^^^^^^^^^^^^
 
-Next we load our Southern Ocean example data which was extracted from the CASCADE database :footcite:p:`deVries2024cascade` and then averaged with depth and time. When applying the pipeline to your own data, note that the data
+Next we create some example data. When applying the pipeline to your own data, note that the data
 needs to be in a `Pandas DataFrame format <https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.html>`_.
 
-In addition to our predictors (`y_train`) we also need environmental data which match our predictions ('X_train'). 
-This data was obtained from monthly climatologies from data sources such as the World Ocean Atlas :footcite:p:`Reagan2024`, NNGv2  :footcite:p:`Broullon2019, Broullon2020`.
-
-.. literalinclude:: ../../../examples/2-phase.py
-   :lines: 29-37
+.. literalinclude:: ../../examples/2-phase.py
+   :lines: 20-23
    :language: python
 
 Training the model
 ^^^^^^^^^^^^^^^^^^
 
 Next we train our model. Note that depending on the number of hyper-parameters specified in the
[…]
