### 2025-09-15 6770f30
tidy assertions
