+++ tests/test.py
@@ -88,13 +88,20 @@ class TestRegressors(unittest.TestCase):
         do_post(statistic="ci95_UL")
         do_post(statistic="ci95_LL")
 
-        integrated_total_poc_mean = pd.read_csv("tests/ModelOutput/regressor/posts/integrated_totals/ens_integrated_totals_mean_poc.csv")['total'][0]
 
-        print("poc mean=")
-        print(integrated_total_poc_mean)
-        
+        integrated_total_poc_mean = pd.read_csv(
+            os.path.join("tests", "ModelOutput", "regressor", "posts", "integrated_totals", "ens_integrated_totals_mean_poc.csv")
+        )["total"].iat[0]
+        integrated_total_poc_ci95_UL = pd.read_csv(
+            os.path.join("tests", "ModelOutput", "regressor", "posts", "integrated_totals", "ens_integrated_totals_ci95_UL_poc.csv")
+        )["total"].iat[0]
+        integrated_total_poc_ci95_LL = pd.read_csv(
+            os.path.join("tests", "ModelOutput", "regressor", "posts", "integrated_totals", "ens_integrated_totals_ci95_LL_poc.csv")
+        )["total"].iat[0]
+
         self.assertAlmostEqual(integrated_total_poc_mean, 1.013e17, delta=1e15)
-
+        self.assertAlmostEqual(integrated_total_poc_ci95_UL, 1.95e17, delta=1e15)
+        self.assertAlmostEqual(integrated_total_poc_ci95_LL, 2.46e16, delta=1e15)
 
 
 class Test2Phase(unittest.TestCase):
@@ -172,21 +179,19 @@ class Test2Phase(unittest.TestCase):
         do_post(statistic="ci95_UL")
         do_post(statistic="ci95_LL")
 
-        integrated_total_poc_mean = pd.read_csv("tests/ModelOutput/2-phase/posts/integrated_totals/ens_integrated_totals_mean_poc.csv")['total'][0]
-
-        print("poc mean=")
-        print(integrated_total_poc_mean)
-        
+        integrated_total_poc_mean = pd.read_csv(
+            os.path.join("tests", "ModelOutput", "2-phase", "posts", "integrated_totals", "ens_integrated_totals_mean_poc.csv")
+        )["total"].iat[0]
+        integrated_total_poc_ci95_UL = pd.read_csv(
+            os.path.join("tests", "ModelOutput", "2-phase", "posts", "integrated_totals", "ens_integrated_totals_ci95_UL_poc.csv")
+        )["total"].iat[0]
+        integrated_total_poc_ci95_LL = pd.read_csv(
+            os.path.join("tests", "ModelOutput", "2-phase", "posts", "integrated_totals", "ens_integrated_totals_ci95_LL_poc.csv")
+        )["total"].iat[0]
+    
         self.assertAlmostEqual(integrated_total_poc_mean, 2.55e17, delta=1e15)
-
-
-        integrated_total_poc_ci95_LL = pd.read_csv("tests/ModelOutput/2-phase/posts/integrated_totals/ens_integrated_totals_ci95_LL_poc.csv")['total'][0]
-
-
-        print("temp total=")
-        print(integrated_total_poc_ci95_LL)
-        
-        self.assertAlmostEqual(integrated_total_poc_ci95_LL, 1.44e+17, delta=1e15)
+        self.assertAlmostEqual(integrated_total_poc_ci95_LL, 3.92e+17, delta=1e15)
+        self.assertAlmostEqual(integrated_total_poc_ci95_UL, 1.44e+17, delta=1e15)
 
 
 
@@ -291,7 +296,7 @@ if __name__ == '__main__':
     # Create a test suite combining all test cases in order
[…]
