### 2025-09-15 57dc841
fix assertion order
