### 2025-09-17 d9f69fb
update entry with doi
