+++ paper/paper.bib
@@ -240,9 +240,14 @@ version = {latest}
   doi={10.1111/j.1600-0587.2008.05742.x},
   publisher={Wiley Online Library}
 }
-@article{devries:2025,
-  title        = {A diverse community constitutes global coccolithophore calcium carbonate stocks},
-  author       = {{de Vries}, Joost and Monteiro, Fanny M. and Poulton, Alex J. and Wiseman, Nicola A. and Wolf, Levi J.},
-  journal      = {bioRxiv},
-  year         = {2025},
+@article {@devries:2025,
+	author = {{de Vries}, Joost and Monteiro, Fanny M. and Poulton, Alex J. and Wiseman, Nicola A. and Wolf, Levi J.},
+	title = {A diverse community constitutes global coccolithophore calcium carbonate stocks},
+	elocation-id = {2025.09.11.675535},
+	year = {2025},
+	doi = {10.1101/2025.09.11.675535},
+	publisher = {Cold Spring Harbor Laboratory},
+	URL = {https://www.biorxiv.org/content/early/2025/09/16/2025.09.11.675535},
+	eprint = {https://www.biorxiv.org/content/early/2025/09/16/2025.09.11.675535.full.pdf},
+	journal = {bioRxiv}
 }
