### 2025-09-17 a09ac9a
tidy text
