+++ paper/paper.bib
@@ -240,7 +240,7 @@ version = {latest}
   doi={10.1111/j.1600-0587.2008.05742.x},
   publisher={Wiley Online Library}
 }
-@article {@devries:2025,
+@article {devries:2025,
 	author = {{de Vries}, Joost and Monteiro, Fanny M. and Poulton, Alex J. and Wiseman, Nicola A. and Wolf, Levi J.},
 	title = {A diverse community constitutes global coccolithophore calcium carbonate stocks},
 	elocation-id = {2025.09.11.675535},
