### 2025-09-17 f1e000c
Update paper.md
Rephrased one sentence
