### 2025-09-17 57b87bb
Add matplotlib seaborn Cartopy sphinx pydata-sphinx-theme sphinx-design sphinxcontrib-bibtex to requirements.txt so that examples can be run locally
