### 2025-09-17 fe1d259
add missing bibliography to BATS example
