+++ requirements.txt
@@ -8,10 +8,3 @@ xgboost==1.7.6
 scikit-learn==1.3.0
 scipy==1.13.1
 scikit-bio==0.6.2
-matplotlib
-seaborn
-Cartopy
-sphinx
-pydata-sphinx-theme
-sphinx-design
-sphinxcontrib-bibtex
