### 2025-09-17 f398949
update deploy.yml to install documentation_requirements from .txt file
