### 2025-09-18 4c03515
remove logs from inside of parallel jobs
