+++ abil/unified_tree_or_bag.py
@@ -1,7 +1,5 @@
 import numpy as np
 import pandas as pd
-import logging
-logger = logging.getLogger("abil")
 
 from sklearn.metrics import mean_absolute_error, balanced_accuracy_score
 
@@ -82,12 +80,12 @@ def estimate_prediction_quantiles(
             threshold=threshold
         )
         regressor_stats = estimate_prediction_quantiles(
-            model.regressor_, 
-            X_predict=X_predict, 
-            X_train=X_train, 
-            y_train=original_y_train, 
-            cv=cv, 
-            chunksize=chunksize, 
+            model.regressor_,
+            X_predict=X_predict,
+            X_train=X_train,
+            y_train=original_y_train,
+            cv=cv,
+            chunksize=chunksize,
             threshold=threshold
         )
         return {
@@ -112,7 +110,7 @@ def estimate_prediction_quantiles(
 
         y_inverse_transformer = getattr(
             model, "inverse_func", FunctionTransformer().inverse_func
-        )       
+        )
 
         model = model.regressor.named_steps["estimator"]
     else:
@@ -133,7 +131,7 @@ def estimate_prediction_quantiles(
         proba = False
     else:
         proba = True
-    
+
 
     ############################################
     # TRANSFORM Y_TRAIN, X_TRAIN AND X_PREDICT #
@@ -143,7 +141,7 @@ def estimate_prediction_quantiles(
         X_train = pd.DataFrame(
             preprocessor.transform(X_train),
             index = getattr(X_train, "index", np.arange(X_train.shape[0])),
-            columns=X_train.columns  
+            columns=X_train.columns
         )
     else:
         raise ValueError("X_train cannot be None")
@@ -151,26 +149,26 @@ def estimate_prediction_quantiles(
         X_predict = pd.DataFrame(
             preprocessor.transform(X_predict),
             index = getattr(X_predict, "index", np.arange(X_predict.shape[0])),
-            columns=X_predict.columns  
+            columns=X_predict.columns
         )
     else:
         raise ValueError("X_predict cannot be None")
 
     if y_train is not None:
         if not proba:
-            y_train = pd.Series(                    
+            y_train = pd.Series(
                 y_transformer(y_train),
                 index = getattr(y_train, "index", np.arange(y_train.shape[0]))
             )
-          
-        else: 
+
+        else:
             y_train = pd.Series(
                 y_train,
                 index = getattr(y_train, "index", np.arange(y_train.shape[0]))
             )
     else:
         raise ValueError("y_train cannot be None")
-    
+
     ##########################################
     # REFIT MODELS WITH TRANSFORMED Y AND Xs #
     ##########################################
@@ -218,12 +216,12 @@ def estimate_prediction_quantiles(
     #################################################
 
     predict_summary_stats = _summarize_predictions(
-        model, 
-        y_inverse_transformer, 
-        X_predict=X_predict, 
+        model,
+        y_inverse_transformer,
+        X_predict=X_predict,
         X_train=X_train,
         y_train=y_train,
-        chunksize=chunksize, 
+        chunksize=chunksize,
         threshold=threshold,
         proba=proba
     )
@@ -267,16 +265,16 @@ def _summarize_predictions(model, y_inverse_transformer, X_predict, X_train=None
             model = BaggingClassifier(estimator=model, n_estimators=160)
             model.fit(X_train, y_train)
         train_pred_jobs = _setup_pred_jobs(
-            model, 
-            X_train, 
-            proba, 
+            model,
+            X_train,
+            proba,
             threshold
         )
         train_results = engine(train_pred_jobs)
 
         if proba:
             losses = 1 - np.array(
-                [balanced_accuracy_score(y_train.astype(int), 
[…]
