### 2025-09-18 6d43bea
switch predict 'multiprocessing' to 'loky'
