### 2025-09-18 7839cb0
Rename classes in JOSS paper
