+++ docs/examples/regressor.py
@@ -4,9 +4,9 @@
 import numpy as np
 from yaml import load
 from yaml import CLoader as Loader
-from abil.tune import tune
-from abil.predict import predict
-from abil.post import post
+from abil.tune import ModelTuner as tune
+from abil.predict import ModelPredictor as predict
+from abil.post import AbilPostProcessor as post
 from abil.utils import example_data 
 
 import os
