### 2025-09-18 c7a0af9
Delete docs/examples/2-phase.py
