### 2025-09-18 f32c790
Delete docs/examples/regressor.py
