### 2025-09-18 da4fd55
Update regressor.py
update class names in import
