### 2025-09-18 f853198
Update regressor.py
Update name to regressor from 2-phase
