+++ examples/regressor.py
@@ -1,5 +1,5 @@
 """
-2-phase Ensemble 
+Regressor Ensemble 
 """
 #paths:
 import os
