+++ examples/2-phase.py
@@ -1,24 +1,24 @@
 """
 2-phase Ensemble 
 """
+#paths:
+import os
 #handling data:
 import numpy as np
 import pandas as pd
 import xarray as xr
 from yaml import load
 from yaml import CLoader as Loader
-#plotting:
-import matplotlib.pyplot as plt
-import cartopy.crs as ccrs
-import cartopy.feature as cfeature
-from matplotlib.colors import Normalize
 #abil functions:
 from abil.tune import ModelTuner as tune
 from abil.predict import ModelPredictor as predict
 from abil.post import AbilPostProcessor as post
 from abil.utils import example_data 
-#paths:
-import os
+#plotting:
+import matplotlib.pyplot as plt
+import cartopy.crs as ccrs
+import cartopy.feature as cfeature
+from matplotlib.colors import Normalize
 
 os.chdir(os.path.join(".", "examples"))
 
