### 2025-09-23 8667a0d
update install instructions
