### 2025-09-23 1671cdc
add pip update instructions
