### 2025-09-30 #8755 @althonos
@nanophyto : Thank you for addressing most of the issues!

Regarding the reproducibility, I followed the "Coccolithopore" example and it worked smoothly (including plots), but I still got different RMSE values, despite running the code in a fresh Python 3.10 environment with the pinned dependencies:
```
error metrics:
R2 = 0.20
relative RMSE = 156.20 (%)
relative MAE = 121.10 (%)
```
where the documentation reports:
```
error metrics:
R2 = 0.09
relative RMSE = 159.22 (%)
relative MAE = 123.73 (%)
```
Is there perhaps some seeding command that is missing somewhere to ensure reproducibility?
