### 2025-10-03 #8755 @nanophyto
> [@nanophyto](https://github.com/nanophyto) : Thank you for addressing most of the issues!
>
> Regarding the reproducibility, I followed the "Coccolithopore" example and it worked smoothly (including plots), but I still got different RMSE values, despite running the code in a fresh Python 3.10 environment with the pinned dependencies:
>
> ```
> error metrics:
> R2 = 0.20
> relative RMSE = 156.20 (%)
> relative MAE = 121.10 (%)
> ```
>
> where the documentation reports:
>
> ```
> error metrics:
> R2 = 0.09
> relative RMSE = 159.22 (%)
> relative MAE = 123.73 (%)
> ```
>
> Is there perhaps some seeding command that is missing somewhere to ensure reproducibility?

Thanks for having another close look at this. It seems that this was an issue with the documentation not being correctly updated after fixing reproducibility issues. The numbers your reported match our numbers, and  have now been updated in PR: https://github.com/nanophyto/Abil/pull/290
