### 2025-10-06 nanophyto/Abil#291 @althonos
[JOSS] Improve management of the `loky` temporary folder
*This issue is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8755*

In `predict.py`, the top of the file reads:

https://github.com/nanophyto/Abil/blob/8c28cb82818244e22ac094101ebbee3bbc8fa9e2/abil/predict.py#L16-L20

Which aims at setting up a temporary folder for the prediction to be used with `loky`. This has two issues:

First, overriding the environment variable which may be user-provided (including potential important configuration if running on a HPC cluster, to avoid setting the temporary folder to a networked filesystem) is an anti-pattern I described in #271.

Second and most importantly: the current code always assumes the current folder can be written to and will attempt to create a temporary folder there. This is not always the case, and could fail. Worse, if the current folder _does_ contain a `tmp` folder, no error will be raised when attempting to create it again (because of the `exist_ok=True`). But later, when calling `shutil.rmtree(temp_folder, ignore_errors=True)` (in `export_prediction`), the folder (which may have already existed, and may have contained non-Abil related data) will get removed with all its contents. **You really don't want to call `rmtree` on a folder you may not have created!**

In my opinion, Abil should just let the users handle the temporary `loky` folder and data. Otherwise, it should try to do so with the `temp_folder` argument of [`Parallel` or `parallel_config`](https://joblib.readthedocs.io/en/latest/parallel.html) (I'm not knowledgeable enough with `joblib` to know which one you want).
