### 2025-10-06 nanophyto/Abil#294 @althonos
[JOSS] Miscellaneous paper remarks
*This issue is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8755*

## References

In **Statement of Need**:

> `Abil` is a Python package to interpolate sparse observations using ensemble-based machine learning algorithms. Oceanographic data is sparse in terms of spatial and temporal distribution due to the nature of collection during oceanographic cruises and requires a more informed approach than traditional gap-filling interpolation. Previous studies have utilized machine-learning methods as a solution to this problem, but often underlying code is difficult to access and deploy. To improve reproducibility within the community, `Abil` was developed as an open-source, user-friendly interface through which machine learning methods can be more easily implemented.

Could you add a reference to one of the "previous studies" you are mentioning?

> In addition, Abil includes essential tools such as pre-implemented pipelines which include environmental feature scaling – a step which is required for algorithms such as nearest neighbor algorithms, optional predictor log transformation – a step desirable in cases where high-value outliers can skew predictions, and zero-stratified cross validation for predictors where absences are more common than occurrences (“zero-inflation”, which is common for ocean biogeochemical observations)

Could you add a reference which perhaps discusses zero-inflated data? E.g. a review?

> For species distribution modelling, post contains estimate_carbon (estimates carbon content
for each target based on value from targets file), def_groups (sums data for groups defined in
a specified dictionary), cwm (calculates the community weighted mean for a specified target),
diversity (calculates the Shannon diversity index) and total (sums all targets, useful for
calculating total species abundances)

Could you add a reference for the Shannon diversity index? It may not be obvious for everybody. Either the original Bell Labs paper or a review/reference paper would be fine in my view.

## Paper sections

Past the first paragraph of the **Statement of Need** I think you are already detailing the method/ results of the Abil package more than talking about the statement of need. I think the manuscript would gain in clarity having several subsequent sections. JOSS does not really enforce a structure besides the presence of **Statement of Need**, but I think you could have a **Result** or **Method** section to mark the content of the next paragraphs.

## Software references

Given that now Abil features pinned software dependencies, I think you should cite the exact version used, e.g. Pandas v2.2.2 (https://doi.org/10.5281/zenodo.10957263) rather than the latest version.
