### 2025-10-06 #8755 @althonos
Hi @ljwolf @nicola-wiseman @nanophyto,

Having read the paper and had a look at the code I think it is now in a much better state and meets the community standard, thanks to your efforts. Apart from one issue ([nanophyto/Abil#291](https://github.com/nanophyto/Abil/issues/291)) and some suggestions for the paper ([nanophyto/Abil#294](https://github.com/nanophyto/Abil/issues/294)) I think I have no more comments and I'm happy to see paper accepted.

Thanks again for your patience and great work!
