### 2025-10-06 bc7e24f
Change the working of the previous studies sentence and add three examples (nitrogen fixation, methane, nitrous oxide).
