+++ paper/paper.bib
@@ -251,3 +251,47 @@ version = {latest}
 	eprint = {https://www.biorxiv.org/content/early/2025/09/16/2025.09.11.675535.full.pdf},
 	journal = {bioRxiv}
 }
+
+@article{luo:2014,
+  AUTHOR = {Luo, Y.-W. and Lima, I. D. and Karl, D. M. and Deutsch, C. A. and Doney, S. C.},
+  TITLE = {Data-based assessment of environmental controls on global marine nitrogen fixation},
+  JOURNAL = {Biogeosciences},
+  VOLUME = {11},
+  YEAR = {2014},
+  NUMBER = {3},
+  PAGES = {691--708},
+  URL = {https://bg.copernicus.org/articles/11/691/2014/},
+  DOI = {10.5194/bg-11-691-2014}
+  }
+
+@article{weber:2019,
+  author={Weber, Thomas
+  and Wiseman, Nicola A.
+  and Kock, Annette},
+  title={Global ocean methane emissions dominated by shallow coastal waters},
+  journal={Nature Communications},
+  year={2019},
+  month={Oct},
+  day={08},
+  volume={10},
+  number={1},
+  pages={4584},
+  abstract={Oceanic emissions represent a highly uncertain term in the natural atmospheric methane (CH4) budget, due to the sparse sampling of dissolved CH4 in the marine environment. Here we overcome this limitation by training machine-learning models to map the surface distribution of methane disequilibrium (∆CH4). Our approach yields a global diffusive CH4 flux of 2--6TgCH4yr−1 from the ocean to the atmosphere, after propagating uncertainties in ∆CH4 and gas transfer velocity. Combined with constraints on bubble-driven ebullitive fluxes, we place total oceanic CH4 emissions between 6--12TgCH4yr−1, narrowing the range adopted by recent atmospheric budgets (5--25TgCH4yr−1) by a factor of three. The global flux is dominated by shallow near-shore environments, where CH4 released from the seafloor can escape to the atmosphere before oxidation. In the open ocean, our models reveal a significant relationship between ∆CH4 and primary production that is consistent with hypothesized pathways of in situ methane production during organic matter cycling.},
+  issn={2041-1723},
+  doi={10.1038/s41467-019-12541-7},
+  url={https://doi.org/10.1038/s41467-019-12541-7}
+  }
+
+@article{yang:2020,
+  author = {Simon Yang  and Bonnie X. Chang  and Mark J. Warner  and Thomas S. Weber  and Annie M. Bourbonnais  and Alyson E. Santoro  and Annette Kock  and Rolf E. Sonnerup  and John L. Bullister  and Samuel T. Wilson  and Daniele Bianchi },
+  title = {Global reconstruction reduces the uncertainty of oceanic nitrous oxide emissions and reveals a vigorous seasonal cycle},
+  journal = {Proceedings of the National Academy of Sciences},
+  volume = {117},
+  number = {22},
+  pages = {11954-11960},
+  year = {2020},
+  doi = {10.1073/pnas.1921914117},
+  URL = {https://www.pnas.org/doi/abs/10.1073/pnas.1921914117},
+  eprint = {https://www.pnas.org/doi/pdf/10.1073/pnas.1921914117},
[…]
