### 2025-10-06 665951a
add shannon 1948 citation
