+++ paper/paper.bib
@@ -262,7 +262,7 @@ version = {latest}
   PAGES = {691--708},
   URL = {https://bg.copernicus.org/articles/11/691/2014/},
   DOI = {10.5194/bg-11-691-2014}
-  }
+}
 
 @article{weber:2019,
   author={Weber, Thomas
@@ -280,7 +280,7 @@ version = {latest}
   issn={2041-1723},
   doi={10.1038/s41467-019-12541-7},
   url={https://doi.org/10.1038/s41467-019-12541-7}
-  }
+}
 
 @article{yang:2020,
   author = {Simon Yang  and Bonnie X. Chang  and Mark J. Warner  and Thomas S. Weber  and Annie M. Bourbonnais  and Alyson E. Santoro  and Annette Kock  and Rolf E. Sonnerup  and John L. Bullister  and Samuel T. Wilson  and Daniele Bianchi },
@@ -294,7 +294,7 @@ version = {latest}
   URL = {https://www.pnas.org/doi/abs/10.1073/pnas.1921914117},
   eprint = {https://www.pnas.org/doi/pdf/10.1073/pnas.1921914117},
   abstract = {N2O is a potent greenhouse gas whose oceanic emissions are still poorly constrained. Here, we reconstruct oceanic N2O emissions from the largest observational synthesis to date, substantially reducing the uncertainty compared to previous estimates and enabling a tightening of the global budget of this gas. We also reveal a vigorous global seasonal cycle, dominated by productive, low-oxygen tropical regions, which suggests sensitivity of the N2O cycle to natural climate variability and anthropogenic climate change. Assessment of the global budget of the greenhouse gas nitrous oxide (N2O) is limited by poor knowledge of the oceanic N2O flux to the atmosphere, of which the magnitude, spatial distribution, and temporal variability remain highly uncertain. Here, we reconstruct climatological N2O emissions from the ocean by training a supervised learning algorithm with over 158,000 N2O measurements from the surface ocean—the largest synthesis to date. The reconstruction captures observed latitudinal gradients and coastal hot spots of N2O flux and reveals a vigorous global seasonal cycle. We estimate an annual mean N2O flux of 4.2 ± 1.0 Tg N⋅y−1, 64\% of which occurs in the tropics, and 20\% in coastal upwelling systems that occupy less than 3\% of the ocean area. This N2O flux ranges from a low of 3.3 ± 1.3 Tg N⋅y−1 in the boreal spring to a high of 5.5 ± 2.0 Tg N⋅y−1 in the boreal summer. Much of the seasonal variations in global N2O emissions can be traced to seasonal upwelling in the tropical ocean and winter mixing in the Southern Ocean. The dominant contribution to seasonality by productive, low-oxygen tropical upwelling systems (\&gt;75\%) suggests a sensitivity of the global N2O flux to El Niño–Southern Oscillation and anthropogenic stratification of the low latitude ocean. This ocean flux estimate is consistent with the range adopted by the Intergovernmental Panel on Climate Change, but reduces its uncertainty by more than fivefold, enabling more precise determination of other terms in the atmospheric N2O budget.}
-  }
+}
 
 @article{nolan:2022,
   author = {Nolan, Victoria and Gilbert, Francis and Reader, Tom},
@@ -304,9 +304,22 @@ version = {latest}
   number = {1},
   pages = {215-232},
   keywords = {conservation, presence–absence, presence-only, sampling bias, species abundance, species distribution model, species occurrence, zero inflation},
-  doi = {https://doi.org/10.1111/jbi.14268},
+  doi = {10.1111/jbi.14268},
   url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jbi.14268},
   eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/jbi.14268},
[…]
