### 2025-10-06 3b93599
additional paper sections
