### 2025-10-06 8bf47c6
add versions for pandas and scikit-bio as they are software specific citations
