+++ paper/paper.bib
@@ -89,7 +89,7 @@ year = {2021}
     title        = {pandas-dev/pandas: Pandas},
     year         = 2025,
     publisher    = {Zenodo},
-    version      = {latest},
+    version      = {2.2.2},
     doi          = {10.5281/zenodo.3509134},
     url          = {https://doi.org/10.5281/zenodo.3509134}
 }
@@ -162,7 +162,7 @@ year = {2021}
   month        = jan,
   year         = 2025,
   publisher    = {Zenodo},
-  version      = {0.6.3},
+  version      = {0.6.2},
   doi          = {10.5281/zenodo.14640761},
   url          = {https://doi.org/10.5281/zenodo.14640761},
   swhid        = {swh:1:dir:3d7e418ed20f8d5c588d18a83c466ee2fb0da6eb
