### 2025-10-06 8b26cfa
fix Nolan 2022 in text citation
