+++ examples/tune_KNN.sh
@@ -11,6 +11,12 @@ i=${SLURM_ARRAY_TASK_ID}
 
 module  load apptainer/1.3.1 
 
+# Define new (unique) temp dir
+JOB_TMP="/user/work/${USER}/.tmp/abil-${SLURM_JOB_ID:-$$}"
+mkdir -p "$JOB_TMP"
+export TMPDIR="$JOB_TMP"
+trap 'rm -rf "$JOB_TMP"' EXIT
+
 singularity exec \
 -B/user/work/$(whoami):/user/work/$(whoami) \
 /user/work/$(whoami)/Abil/singularity/abil.sif \
