+++ examples/post.sh
@@ -6,13 +6,13 @@
 #SBATCH --mem=60000M
 #SBATCH --cpus-per-task=1
 
-module  load apptainer/1.3.1 
-
 # Define new (unique) temp dir
 JOB_TMP="/user/work/${USER}/.tmp/abil-${SLURM_JOB_ID:-$$}"
 mkdir -p "$JOB_TMP"
 export TMPDIR="$JOB_TMP"
 trap 'rm -rf "$JOB_TMP"' EXIT
+
+module  load apptainer/1.3.1 
 
 singularity exec \
 -B/user/work/$(whoami):/user/work/$(whoami) \
