### 2025-10-08 nanophyto/Abil#291 @nanophyto
Thanks a lot for this feedback. Removing the entire (non-unique) tmp folder was a big oversight.

I have now removed this from the package, and have updated the HPC docs with instructions on how to define the loky tmp folder on the user side as part of the SLURM bash scripts:

```
# Define new (unique) temp dir
JOB_TMP="/user/work/${USER}/.tmp/abil-${SLURM_JOB_ID:-$$}"
mkdir -p "$JOB_TMP"
export TMPDIR="$JOB_TMP"
trap 'rm -rf "$JOB_TMP"' EXIT
```

https://github.com/nanophyto/Abil/pull/296
