### 2025-10-08 nanophyto/Abil#291 @althonos
I mean, I think in SLURM the `$TMPDIR` variable is almost always set by the sysadmin to some sensible temporary folder, so I would not necessary modify it. However, you could just create the Loky folder in there:

```bash
LOKY_TEMP_FOLDER="$TMPDIR/abil-${SLURM_JOB_ID:-$$}"
mkdir -p "$LOKY_TEMP_FOLDER"
abil ...
rm -rd $LOKY_TEMP_FOLDER
```
