### 2025-10-08 nanophyto/Abil#291 @nanophyto
> I mean, I think in SLURM the `$TMPDIR` variable is almost always set by the sysadmin to some sensible temporary folder, so I would not necessary modify it. However, you could just create the Loky folder in there:
>
> LOKY_TEMP_FOLDER="$TMPDIR/abil-${SLURM_JOB_ID:-$$}"
> mkdir -p "$LOKY_TEMP_FOLDER"
> abil ...
> rm -rd $LOKY_TEMP_FOLDER

Yes that would be reasonable! But I guess this is quite a cluster specific thing. At the cluster we use, we do not have write access to a temp folder only e.g. `/user/work/username` and  `/user/home/username`. Thus the creation of a temp folder within the user directory.

However, we explicitly state in the docs that this is cluster specific, and it is a bit beyond the scope of the package itself. So I will leave the suggestion as is.
