### 2025-10-08 60cc343
Update paper/paper.bib
