+++ paper/paper.bib
@@ -90,7 +90,7 @@ year = {2021}
     year         = 2025,
     publisher    = {Zenodo},
     version      = {2.2.2},
-    doi          = {10.5281/zenodo.3509134},
+    doi          = {10.5281/zenodo.10957263},
     url          = {https://doi.org/10.5281/zenodo.3509134}
 }
 
