### 2025-10-08 2c527ba
Update paper/paper.bib
