+++ paper/paper.bib
@@ -163,7 +163,7 @@ year = {2021}
   year         = 2025,
   publisher    = {Zenodo},
   version      = {0.6.2},
-  doi          = {10.5281/zenodo.14640761},
+  doi          = {10.5281/zenodo.593387},
   url          = {https://doi.org/10.5281/zenodo.14640761},
   swhid        = {swh:1:dir:3d7e418ed20f8d5c588d18a83c466ee2fb0da6eb
                    ;origin=https://doi.org/10.5281/zenodo.593387;visi
