### 2025-10-08 2e5d472
Update paper/paper.bib
