+++ paper/paper.bib
@@ -91,7 +91,7 @@ year = {2021}
     publisher    = {Zenodo},
     version      = {2.2.2},
     doi          = {10.5281/zenodo.10957263},
-    url          = {https://doi.org/10.5281/zenodo.3509134}
+    url          = {https://doi.org/10.5281/zenodo.10957263}
 }
 
 @InProceedings{mckinney:2010,
