### 2025-10-08 78aba81
Update paper/paper.bib
