+++ paper/paper.bib
@@ -164,7 +164,7 @@ year = {2021}
   publisher    = {Zenodo},
   version      = {0.6.2},
   doi          = {10.5281/zenodo.593387},
-  url          = {https://doi.org/10.5281/zenodo.14640761},
+  url          = {https://doi.org/10.5281/zenodo.593387},
   swhid        = {swh:1:dir:3d7e418ed20f8d5c588d18a83c466ee2fb0da6eb
                    ;origin=https://doi.org/10.5281/zenodo.593387;visi
                    t=swh:1:snp:91a2b0adb7ab0ff47a8367fcd810c0c0487302
