### 2025-10-08 #8755 @nanophyto
@crvernon I believe we have now implemented all requested changes (including some new ones by @althonos above).

Please let us know how to proceed as we are all keen to get a DOI on this for e.g. grant proposals.

Thanks!
