### 2025-10-14 9402c1c
fix inline Meyer ref
