### 2025-10-20 3cda356
expand docstrings
