### 2025-10-20 d9ff5ab
rename function approximate_Gamma to generate_gamma_samples
