### 2025-10-20 0da28dc
add minimal example
