+++ abil/analyze.py
@@ -216,7 +216,7 @@ def area_of_applicability(
     if return_all:
         return aoa_, di_test_, lpd_test_, cutpoint, test_to_train_d_
     else:
-        return aoa, di_test, cutpoint
+        return aoa_, di_test_, cutpoint
 
 if __name__ == "__main__":
     import pandas as pd
