### 2025-11-28 50719f1
add logging configuration to tune, predict, and post so logger.info exports to slurm.out
