### 2025-12-01 cc5df38
add logging configuration to tune, predict, and post so logger.info exports to slurm.out (#302)
