### 2025-12-01 2a4348f
add best fit log or nolog print after regressor fitting
