### 2025-12-02 4294566
update log_grid_search to export best_transformation as part of best_estimator
