+++ abil/log_grid_search.py
@@ -123,6 +123,8 @@ class LogGridSearch:
             grid_search = GridSearchCV(model, param_grid = self.param_grid, scoring=self.scoring, refit=True,
                             cv = self.cv, verbose = self.verbose, return_train_score=True, error_score=-1e99)
             grid_search.fit(X, y)
+            best_transformation = "log"
+
 
         elif log == "no":
 
@@ -130,6 +132,8 @@ class LogGridSearch:
             grid_search = GridSearchCV(model, param_grid = self.param_grid, scoring=self.scoring, refit=True,
                             cv = self.cv, verbose = self.verbose, return_train_score=True, error_score=-1e99)
             grid_search.fit(X, y)
+            best_transformation = "nolog"
+
 
         elif log == "both":
 
@@ -153,6 +157,9 @@ class LogGridSearch:
             else:
                 grid_search = grid_search1
                 best_transformation = "nobest"
+                
             grid_search.cv_results_['best_transformation'] = best_transformation
+
+        grid_search.best_estimator_.best_transformation_ = best_transformation
 
         return grid_search
