### 2025-12-02 287ad04
add transformation_ argument to best_estimator_ for access in post
