### 2025-12-02 6ea61cf
print reg_grid_search.best_estimator_.transformation_ to slurm logger
