+++ abil/tune.py
@@ -256,7 +256,7 @@ class ModelTuner:
                 reg_grid_search = reg.transformed_fit(X_train, y, log, self.model_config['predictors'].copy())
             
             if log == 'both':
-                logger.info(f"best fit: {reg_grid_search.cv_results_['best_transformation']}")
+                logger.info(f"best fit: {reg_grid_search.best_estimator_.transformation_}")
             m2 = reg_grid_search.best_estimator_
 
 
