### 2025-12-02 7d670ca
add transformation to post parameters merge for regressor and zir
