+++ abil/post.py
@@ -334,8 +334,9 @@ class AbilPostProcessor:
                     max_samples = m.regressor_.named_steps.estimator.max_samples
                     min_samples_leaf = m.regressor_.named_steps.estimator.min_samples_leaf
                     n_estimators = m.regressor_.named_steps.estimator.n_estimators
+                    transformation = m.transformation_
                     parameters = pd.DataFrame({'target':[target], 'n_estimators':[n_estimators], 'max_features':[max_features], 'max_depth':[max_depth], 
-                                            'min_samples_leaf':[min_samples_leaf], 'max_samples':[max_samples]
+                                            'min_samples_leaf':[min_samples_leaf], 'max_samples':[max_samples], 'transformation':[transformation]
                                             })
                     all_parameters.append(parameters)
                 elif model == "xgb":
@@ -346,9 +347,10 @@ class AbilPostProcessor:
                     colsample_bytree = m.regressor_.named_steps.estimator.colsample_bytree
                     gamma = m.regressor_.named_steps.estimator.gamma
                     alpha = m.regressor_.named_steps.estimator.reg_alpha
+                    transformation = m.transformation_
                     parameters = pd.DataFrame({'target':[target], 'learning_rate':[learning_rate], 'n_estimators':[n_estimators], 
                                             'max_depth':[max_depth], 'subsample':[subsample], 'colsample_bytree':[colsample_bytree],
-                                            'learning_rate':[learning_rate], 'gamma':[gamma], 'alpha':[alpha]                                           
+                                            'learning_rate':[learning_rate], 'gamma':[gamma], 'alpha':[alpha], 'transformation':[transformation]                                           
                                             })
                     all_parameters.append(parameters)
                 elif model == "knn":
@@ -358,8 +360,9 @@ class AbilPostProcessor:
                     n_neighbors = m.regressor_.named_steps.estimator.estimator.n_neighbors
                     p = m.regressor_.named_steps.estimator.estimator.p
                     weights = m.regressor_.named_steps.estimator.estimator.weights
+                    transformation = m.transformation_
                     parameters = pd.DataFrame({'target':[target], 'max_samples':[max_samples], 'max_features':[max_features],
-                                            'leaf_size':[leaf_size], 'n_neighbors':[n_neighbors], 'p':[p], 'weights':[weights]
+                                            'leaf_size':[leaf_size], 'n_neighbors':[n_neighbors], 'p':[p], 'weights':[weights],'transformation':[transformation]
                                             })
                     all_parameters.append(parameters) 
 
@@ -373,6 +376,7 @@ class AbilPostProcessor:
                     max_samples_reg = m.regressor_.regressor.named_steps.estimator.max_samples
                     min_samples_leaf_reg = m.regressor_.regressor.named_steps.estimator.min_samples_leaf
                     n_estimators_reg = m.regressor_.regressor.named_steps.estimator.n_estimators
+                    transformation = m.transformation_
 
                     n_estimators_clf = m.classifier.named_steps.estimator.n_estimators
                     max_features_clf = m.classifier.named_steps.estimator.max_features
@@ -383,6 +387,7 @@ class AbilPostProcessor:
                     parameters = pd.DataFrame({'target':[target], 'reg_n_estimators':[n_estimators_reg], 
                                             'reg_max_features':[max_features_reg], 'reg_max_depth':[max_depth_reg], 
                                             'reg_min_samples_leaf':[min_samples_leaf_reg], 'reg_max_samples':[max_samples_reg],
+                                            'reg_transformation':[transformation],
[…]
