### 2025-12-02 dda9b3f
fix transformation call for zir
