### 2025-12-02 8397700
copy transformation attribute when cloning regressor
