### 2025-12-02 cb3e010
rewrite logger.info() statement to call correct attribute from reg_grid_search
