### 2025-12-02 2160296
rewrite transformation calls in post for parameter merging
