### 2025-12-02 9522b7c
fix non zir transformation call
