### 2025-12-02 dd116f6
reformat hyperparameter calling for pipeline or bare regressor
