### 2025-12-02 b919157
fix zir to same formating as reg only
