+++ abil/post.py
@@ -413,86 +413,110 @@ class AbilPostProcessor:
                 raise ValueError("classifiers are not supported")
 
             elif self.model_type == "zir":
+                # Extract inner estimators
+                reg_est = get_inner_estimator(m.regressor_.regressor)
+                transformation = get_transformation(m.regressor_.regressor)
+
+                clf_est = get_inner_estimator(m.classifier)
+
                 if model == "rf":
-                    max_depth_reg = m.regressor_.regressor.named_steps.estimator.max_depth
-                    max_features_reg = m.regressor_.regressor.named_steps.estimator.max_features
-                    max_samples_reg = m.regressor_.regressor.named_steps.estimator.max_samples
-                    min_samples_leaf_reg = m.regressor_.regressor.named_steps.estimator.min_samples_leaf
-                    n_estimators_reg = m.regressor_.regressor.named_steps.estimator.n_estimators
-                    transformation = m.regressor_.regressor.named_steps.estimator.transformation_
+                    max_depth_reg = reg_est.max_depth
+                    max_features_reg = reg_est.max_features
+                    max_samples_reg = reg_est.max_samples
+                    min_samples_leaf_reg = reg_est.min_samples_leaf
+                    n_estimators_reg = reg_est.n_estimators
 
-                    n_estimators_clf = m.classifier.named_steps.estimator.n_estimators
-                    max_features_clf = m.classifier.named_steps.estimator.max_features
-                    max_depth_clf = m.classifier.named_steps.estimator.max_depth
-                    min_samples_leaf_clf = m.classifier.named_steps.estimator.min_samples_leaf
-                    max_samples_clf = m.classifier.named_steps.estimator.max_samples
+                    n_estimators_clf = clf_est.n_estimators
+                    max_features_clf = clf_est.max_features
+                    max_depth_clf = clf_est.max_depth
+                    min_samples_leaf_clf = clf_est.min_samples_leaf
+                    max_samples_clf = clf_est.max_samples
 
-                    parameters = pd.DataFrame({'target':[target], 'reg_n_estimators':[n_estimators_reg], 
-                                            'reg_max_features':[max_features_reg], 'reg_max_depth':[max_depth_reg], 
-                                            'reg_min_samples_leaf':[min_samples_leaf_reg], 'reg_max_samples':[max_samples_reg],
-                                            'reg_transformation':[transformation],
-                                            'clf_n_estimators':[n_estimators_clf], 
-                                            'clf_max_features':[max_features_clf], 'clf_max_depth':[max_depth_clf], 
-                                            'clf_min_samples_leaf':[min_samples_leaf_clf], 'clf_max_samples':[max_samples_clf]
-                                            })
+                    parameters = pd.DataFrame({
+                        'target':[target],
+                        'reg_n_estimators':[n_estimators_reg],
+                        'reg_max_features':[max_features_reg],
+                        'reg_max_depth':[max_depth_reg],
+                        'reg_min_samples_leaf':[min_samples_leaf_reg],
+                        'reg_max_samples':[max_samples_reg],
+                        'reg_transformation':[transformation],
+                        'clf_n_estimators':[n_estimators_clf],
+                        'clf_max_features':[max_features_clf],
+                        'clf_max_depth':[max_depth_clf],
+                        'clf_min_samples_leaf':[min_samples_leaf_clf],
+                        'clf_max_samples':[max_samples_clf]
+                    })
                     all_parameters.append(parameters)
 
                 elif model == "xgb":
-                    learning_rate_reg = m.regressor_.regressor.named_steps.estimator.learning_rate
[…]
