### 2025-12-03 4c46544
update test to include "both" for log transformation search
