### 2025-12-03 18ee393
Fix logger in tune for transformation output
