+++ abil/tune.py
@@ -256,7 +256,14 @@ class ModelTuner:
                 reg_grid_search = reg.transformed_fit(X_train, y, log, self.model_config['predictors'].copy())
             
             if log == 'both':
-                logger.info(f"best fit: {reg_grid_search.best_estimator_.named_steps['estimator'].transformation_}")
+                best = reg_grid_search.best_estimator_
+                # If it's a pipeline → extract "estimator"
+                if hasattr(best, "named_steps"):
+                    ttr = best.named_steps["estimator"]
+                else:
+                    # Bare Transformed Target Regressor
+                    ttr = best
+                logger.info(f"best fit: {ttr.transformation_}")
             m2 = reg_grid_search.best_estimator_
 
 
