### 2025-12-03 dbbd4fc
fix assertion numbers with new model config (log transforms)
